### 2018-08-15 #244 @sckott
hi @JonasGeschke Thanks for your submission.

I have a clarifying question. What is the relationship of Species+/CITES to IUCN Red List, if any?  Does it contain any overlapping information/data?

(p.s. you didn't check _Do you intend for this package to go on CRAN?_ - i think you're pkg is already on CRAN, so go ahead and check that)
