### 2018-08-16 #244 @JonasGeschke
Hi @sckott thanks for your quick message.

There is no relationship between the IUCN Red List and the CITES Speciesplus database. The IUCN Red List contains data about the conservation/endangerment status of species; the CITES Speciesplus database contains data about the (illegal) trade status of endangered species, so it goes beyond the Red List. Actually, the Speciesplus database also contains data about the protection status of species under the Convention on Migrating Species (CMS), but the API so far only allows access to the CITES legislation status information.

Regarding the _Do you intend for this package to go on CRAN?_ question thats right. By mistake, we already submitted the package to CRAN before having it reviewed via ROpenSci. So we dont intend the package to go on CRAN, as it already is on CRAN. Thus, we are looking forward to the reviewing process and how big the first update of the package on CRAN will be.
