### 2018-08-16 #244 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @JonasGeschke   !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). If you haven't used `goodpractice` it's an R package that checks a number of things with another package - most of which we agree with and want authors to follow. You don't need to address these now, it's info for reviewers to use to get started. There's very little in the report below, which is a good thing!

```r
── GP rcites ──────────

It is good practice to

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about
    80 characters wide. Try make your lines shorter than 80 characters

    R/sppplus_simplify.R:29:1
    R/sppplus_simplify.R:32:1
    R/sppplus_simplify.R:50:1
    R/sppplus_taxonconcept.R:35:1
    R/sppplus_taxonconcept.R:48:1
    ... and 30 more lines
```

Seeking reviewers now 🕐

---

Reviewers:

- @noamross  due date 2018-09-07
- xxx
