### 2018-08-17 #244 @KevCaz
@sckott thanks!

Actually, following the guidelines of Ropenscience, we've used `goodpractice` (that's why we got rid of a `sapply()`!). Regarding the length of code line issue, I'd be happy to have your feed back: we used `tidy_dir()` from `formatR` package with a `width.cutoff=80`, so, should we reduce the cutoff?
