### 2018-08-20 bd40057
add ROpenSci badge
