### 2018-08-20 #244 @sckott
If those lines are already 80 or shorter, then all good. I didn't check them myself. there are definitely exceptions when it's just too cumbersome to do so
