### 2018-08-21 ccb4bba
Fix top-level package roxygen documentation
