### 2018-08-21 #244 @noamross
-   See comments on README below
  -   No examples for `spplus_simplify()`

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is a very useful package to retrieve species CITES status and history.  I've actually developed a related package: <https://github.com/ecohealthalliance/cites/>, which provides access to data pulled from the non-API friendly trade.cites.org.  I was very glad to see **rcites** and it work together just fine!

The package is well-constructed and documented. Setting up authentication follows best practices and was fast and easy.

The major challenge that the package has to deal with is the complexity of returned objects from the API.  I think this could be better handled in both API and documentation.  I also think the authors should consider how to deal with workflows for bulk analysis. I detail more on these and a few other things below.

#### Simplifying/processing objects

The API returns fairly complex records that are challenging to wrangle, which the authors simplify in most functions, with the option of simplifying further with `simplify=TRUE`  options.  This works OK but should be more aggressive.  Even simplified tables still have, for instance, list entries which are lists of NULL objects.  This might be a matter of including more list types as "special cases" (e.g., `"term"`, `"source"`, `"eu_decision_type"`), or it might require more special handling on a field-by-field basis.

I also note that "special cases" fields that end up as factors when most of them (like URLs) should be characters.

I think aggressive simplification should be the default output, and the alternative should be a simple unprocessed list (maybe `raw=FALSE`), which would let users then deal with post-processing themselves if they have some specialized need.

I note that the functions return `data.tables`, not data frames.  I think this is unnecessary and can even trip people up who are not expecting this data type.  (Follow the _principle of least astonishment_<sup>1</sup>).  The `data.table` package is great but it's primary benefit - performance - is basically nonexistent here when we are dealing with tables of maybe of dozens of rows.

#### Printing objects

Because the objects returned are complex, and also because they often include very long text fields, they don't print well.  There are a couple of ways to approach this. One would be to make the returned objects custom S3 classes (e.g., `"spl_leg"`, `"spl_eu_leg`") with their own print methods that print summaries.  A second would be to assign the data frames in the returned objects class `c("tbl_df", "tbl", "data.frame")`.  This would provide more succinct printing for users who have tidyverse packages loaded without adding those packages as dependencies.

If you return the raw list I suggest above, you could take a tip from the **gh** package, give it the class `c("list", "spl_raw"`) and have a nice JSON print method like [shown here](https://github.com/cran/gh/blob/b51273727796ea545f57a765270dd794eb544c31/R/print.R).

#### Function API

The authors use the `sppplus_` prefix for some functions, and `taxon_` for others. It might be simpler to have all functions retrieving something from the API to start with the same prefix (if you want it shorter, I suggest `spl_`).

I note you use `data.table` to modify tables _in place_ with `spplus_simplify()`.  This is confusing and non-standard behavior.  It would make more sense to return the simplified data frames as objects (e.g. `simpl_df <- sppplus_simplify(res1)`).  The current behavior precludes a functional workflows such as simplifying all tables in an object (`leg_smpl <- lapply(leg1, sppplus_simplify)`).

None of the functions currently have an argument to define the `language` parameter of the API, which I think is important.  Also `taxon_cites/eu_legislation()` functions don't accept arguments for the `scope` parameter, and should.

`sppplus_taxonconcept` accepts a single taxon name and returns a single taxon concept record, but this [API endpoint](http://api.speciesplus.net/documentation/v1/taxon_concepts/index.html) has many more parameters for selecting taxon concepts in bulk.  All these should be made available (see "Bulk Analysis" below), though perhaps you want to break this into two functions - one to grab info on one species, and one for the broader downloads that aggregates the data differently.

#### Documentation, README, and Vignette

The documentation is very good, and I'm glad that everything refers back to the appropriate API documentation page.  I'd suggest just a little more wording in the `@return` fields like, "see the API documentation for definitions of all return fields."

The README is a bit sparse.  It could use more of an introduction, in which the authors describe what the Species+ API _is_ and what data can be retrieved from it, and a few lines of basic usage examples. (See https://ropensci.github.io/dev_guide/building.html#readme).  Remember the _principle of first entry_ <sup>1</sup> - any piece of documentation may be the first encounter the user has with the API, data source, or even the CITES treaty/organization.  Don't pass up the opportunity to give a curious data scientist some environmental education!

The vignette, probably because of the need to make API calls, doesn't actually compile and therefore doesn't show outputs from the code that is run.  (The outputs also make for challenging reading, which is why I suggest better printing methods above).  This isn't very helpful and I suggest that the authors compile the vignette to a markdown output locally and have a vignette with all the evaluated output blocks (still name `*.Rmd*) in the `vignettes/` folder.  This will allow for a useful vignette on CRAN and the **pkgdown** site. (But you need to remember to update it manually!)

While the vignette demonstrates use of a few functions, it would be much more helpful if it explained the type of information that was returned, the structure of the object, and demonstrate a few additional workflow of manipulating these results and using them in analysis, e.g.:

> `taxon_cites_legislation()`" returns information on legislative acts related to the species.  Let's see which countries have quotas on exports of Elephant ivory....The object returned is a list of three types of actions taken - listings, quotas, and suspensions.  Here we extract and work with `cites_quotas`...

The vigenette doesn't show `sppplus_simplify()` or explain anything about the `simplify` arguments.  These are key to understanding the package workflow and should be elaborated on along with explanations of the data structures returned, though of course this depends on changes you may make to the simplification approaches described above.

#### Bulk Analysis

The package currently only has non-vectorized functions and focuses on a workflow where one is examining information on a single species.  I think there is a lot of potential for enabling different analyses.  First, as I mention above, all the taxon concept endpoint parameters should be accessible so that one can retrieve a list of taxa to operate on. A natural extension of this would be to at least show in the vignette how to get the paginated return values, and then iterate over them to retrieve, say, legislation for all those species.  I think a better solution would be to build auto-pagination into the taxon concept function, and then add vectorization into the rest of the functions (thinking carefully of how to aggregate all the results into a set of usable data frames).

A major change to consider would be to cache/store downloaded data in a local database that can then be queried in more ways.  For instance, it would be good to be able to ask "What quotas were started in 2010?".  I note that the Species+ API [anticipates this](https://api.speciesplus.net/nomenclature), describing a workflow where data is cached in the local client and updated by calling the `taxon_concepts` endpoint with the `updated_at` parameter to get a list of species whose information has changed since the last download. A few RO packages have this design pattern, like [**taxizedb**](https://github.com/ropensci/taxizedb), [**restez**](https://github.com/AntonelliLab/restez), and my own **cites** package.  I won't demand this for the current review but think you should consider it for 1.0 or 2.0.

# Tiny stuff

None of the >80 character line widths are egregious or problematic, but I recommend shortening the lines in the examples to this width or less so that they show up help files without overrunning the width of the viewer.

Implemented as a [PR](https://github.com/ibartomeus/rcites/pull/28):

-  Rename vignette for current package name
-  Fix top-level package docs roxygen syntax

---

<sup>1</sup> https://en.wikipedia.org/wiki/Principle_of_least_astonishment

<sup>2</sup> This one I made up a while back, but it's about what we recommend at <https://ropensci.github.io/dev_guide/building.html#readme>
