### 2018-08-22 #244 @JonasGeschke
@noamross thank you for this very comprehensive review. Your comments will help us a lot, we will do our best to address them all! Also I like that your `cites` package for the trade database, which actually was a thought of mine to potentially include this in `rcites` lateron, too.
I will work out a checklist from your comments later today, so we can start implementing them step by step.
