### 2018-08-22 #244 @JonasGeschke
**Review checklist to be implemented**

- [ ] make `simplify` more aggressive and provide option to receive unprocessed list (`raw=FALSE`)
- [ ] check `special_cases` for correct data structure
- [ ] consider returning data.frames rather than data.tables
- [ ] make printing of long text fields more clean
- [ ] rename function prefixes for more straight forward understanding and workflow
- [ ] return simplified data.frames as objects
- [ ] consider a language parameter
- [ ] consider a scope parameter for `cites_ (and) eu_legislation`
- [ ] consider a bulk analysis for `taxonconcept` (+ auto-pagination)
- [ ] add more wording to the @return fields of documentation
- [ ] add introduction and basic information to README
- [ ] make the vignette more comprehensive, including code outputs + explanation of type/structure of information that is returned and workflows + explanation of `simplify`
- [ ] consider caching downloaded data locally for further processing
- [ ] shorten lines in examples

_to be completed from second review_
