### 2018-08-22 b1e3904
small edits associated with ROpenSci review
