### 2018-08-22 f8e2bc3
minor rewording
