### 2018-08-22 #244 @KevCaz
@noamross I have two questions regarding your review:

> functions currently have an argument to define the language parameter of the API

Do you mean JSON/XML?

Also, regarding data.table, do you think we should rather not use it? I think we'll be able to do the same manipulations without it, so it means we'll have one less dependency, which may be desired.

Thanks!
