### 2018-08-22 #244 @noamross
Correct, I meant the English/French/Spanish parameter.

Yes, I'd drop data.table as a dependency.

Also, my advice is to hold off on major changes prior to getting your second review! It's usually easier to address them simultaneously, and @mcsiple may not agree with me on everything!
