### 2018-08-23 318e3bb
minor re-wording
