+++ tests/testthat/test-wrongtoken.R
@@ -2,7 +2,7 @@ context("wong token")
 
 token <- "wrong"
 msg <- "Unauthorized (HTTP 401)"
-q_url <- sppplus_url(paste0("taxon_concepts/", tx_id, "/distributions.json"))
+q_url <- rcites_url(paste0("taxon_concepts/", tx_id, "/distributions.json"))
 
 test_that("expected url", {
   expect_equal(q_url, "https://api.speciesplus.net/api/v1/taxon_concepts/4521/distributions.json")
@@ -10,7 +10,7 @@ test_that("expected url", {
 
 test_that("expected errors", {
   ut_pause()
-  expect_error(sppplus_res(q_url, token), msg, fixed = TRUE)
+  expect_error(rcites_res(q_url, token), msg, fixed = TRUE)
   ut_pause()
-  expect_error(sppplus_taxonconcept(query_taxon = 'Loxodonta africana', token = token), msg, fixed = TRUE)
+  expect_error(spp_taxonconcept(query_taxon = 'Loxodonta africana', token = token), msg, fixed = TRUE)
 })
