### 2018-08-24 3868c9d
SPECIESPLUS_TOKEN changes and logo update
