+++ R/spp_taxonconcept.R
@@ -34,30 +34,30 @@
 
 spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL) {
     # token check
-    if (is.null(token))
+    if (is.null(token)) 
         token <- rcites_getsecret()
     query <- gsub(pattern = " ", replacement = "%20", x = query_taxon)
-    #
+    # 
     q_url <- rcites_url(paste0("taxon_concepts.json", "?name=", query))
     res <- rcites_res(q_url, token)
-
+    
     if (!res$pagination$total_entries) {
         warning("Taxon not listed.")
         out <- NULL
     } else {
-        nm <- c("id", "full_name", "author_year", "rank", "name_status", "updated_at",
+        nm <- c("id", "full_name", "author_year", "rank", "name_status", "updated_at", 
             "active", "cites_listing")
         tmp <- as.data.table(do.call(rbind, lapply(res$taxon_concepts, function(x) rbind(x[nm]))))
-        #
+        # 
         rcites_simplify(tmp)
-        #
+        # 
         if (isTRUE(appendix_only)) {
             out <- tmp
         } else {
             out <- list()
             out$all <- tmp
             tmp <- res$taxon_concepts[[1L]]
-            #
+            # 
             if ("synonyms" %in% names(tmp)) {
                 out$synonyms <- as.data.table(do.call(cbind, tmp$synonyms[[1L]]))
             }
@@ -69,6 +69,6 @@ spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL) {
             }
         }
     }
-    #
+    # 
     out
 }
