### 2018-08-24 6cc9343
rcites_url now uses the dots argument for flexibility
