+++ R/spp_taxonconcept.R
@@ -11,6 +11,7 @@
 #' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
 #' set directly in \code{Renviron}. Alternatively, \code{set_token()} can
 #' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
+#' @param raw a logical. Should raw data be returned?
 #'
 #' @return
 #' If \code{appendix_only} is \code{TRUE}, then a data frame with a species'
@@ -22,7 +23,6 @@
 #' @references
 #' \url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
 #'
-#' @importFrom data.table as.data.table :=
 #' @export
 #'
 #' @examples
@@ -32,41 +32,34 @@
 #' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
 #' }
 
-spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL) {
+spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL, raw = FALSE) {
     # token check
     if (is.null(token)) 
         token <- rcites_getsecret()
     query <- gsub(pattern = " ", replacement = "%20", x = query_taxon)
     # 
     q_url <- rcites_url("taxon_concepts.json", "?name=", query)
-    res <- rcites_res(q_url, token)
-    
-    if (!res$pagination$total_entries) {
+    tmp <- rcites_res(q_url, token)
+    # empty output
+    if (!tmp$pagination$total_entries) {
         warning("Taxon not listed.")
         out <- NULL
     } else {
-        nm <- c("id", "full_name", "author_year", "rank", "name_status", "updated_at", 
-            "active", "cites_listing")
-        tmp <- as.data.table(do.call(rbind, lapply(res$taxon_concepts, function(x) rbind(x[nm]))))
-        # 
-        rcites_simplify(tmp)
-        # 
-        if (isTRUE(appendix_only)) {
-            out <- tmp
+        if (raw) {
+            out <- tmp$taxon_concepts
+            class(out) <- c("list", "spp_raw")
         } else {
-            out <- list()
-            out$all <- tmp
-            tmp <- res$taxon_concepts[[1L]]
-            # 
-            if ("synonyms" %in% names(tmp)) {
-                out$synonyms <- as.data.table(do.call(cbind, tmp$synonyms[[1L]]))
-            }
-            if ("common_names" %in% names(tmp)) {
-                out$common_names <- as.data.table(do.call(cbind, tmp$common_names[[1L]]))
-            }
-            if ("higher_taxa" %in% names(tmp)) {
-                out$higher_taxa <- as.data.table(do.call(cbind, tmp$higher_taxa))
-            }
+            sp_nm <- c("higher_taxa", "common_names", "synonyms", "cites_listings", 
+                "cites_listing", "accepted_names")
+            out <- do.call(rbind, lapply(tmp$taxon_concepts, function(x) unlist(x[!names(x) %in% 
+                sp_nm])))
+            # special cases return extra data frame sp_nm <- c('higher_taxa', 'common_names',
+            # 'synonyms', 'cites_listings', 'cites_listing', 'accepted_names') vc_nm <-
+            # unique(unlist(lapply(tmp2, names))) vc_nm <- vc_nm[! vc_nm %in% sp_nm] out <-
+            # lapply( lapply(tmp2, function(x) x[names(x) %in% vc_nm]), function(x) rbind(x,
+            # setdiff(vc_nm, names(x)) ))
+            class(out) <- c("tbl_df", "tbl", "data.frame", "spp_tax")
+            ## 
         }
     }
     # 
