+++ tests/testthat/test_pkghelpers.R
@@ -6,3 +6,6 @@ test_that("basic functioning", {
   expect_equal(rcites_url("extra"), "https://api.speciesplus.net/api/v1/extra")
   expect_message(set_token(""), "no token provided")
 })
+
+# to be tested 
+# as.data.frame(do.call(rbind, res$taxon_concepts[[1]][[10]]))
