### 2018-08-25 7edbea8
test rcites_numberpages() :shield:
