+++ R/zzz.R
@@ -51,7 +51,7 @@ rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 
 # pagination
 rcites_numberpages <- function(x) {
-    x$total_entries%/%x$per_page + (x$total_entries%/%x$per_page > 0)
+    x$total_entries%/%x$per_page + (x$total_entries%%x$per_page > 0)
 }
 
 
