+++ man/spp_taxonconcept.Rd
@@ -36,6 +36,7 @@ Retrieve the taxon concept of a specific taxon (scientific name).
 res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
 res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', appendix_only = TRUE)
 res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
+res4 <- spp_taxonconcept(query_taxon = 'Amazilia')
 }
 }
 \references{
