+++ R/spp_cites_legislation.R
@@ -5,7 +5,15 @@
 #'
 #' @param taxon_id character string containing a species' taxon concept identifier
 #' (see \code{\link[rcites]{spp_taxonconcept}}).
-#' @param type vector of character strings indicating type of legislation information requested, values are taken among \code{listing}, \code{quota} and \code{suspension}. Default includes the three of them.
+#' @param type vector of character strings indicating the type of legislation 
+#' information requested, values are taken among \code{listings}, \code{quotas} and 
+#' \code{suspensions}. Default includes the three of them.
+#' @param scope vector of character strings indicating the time scope of legislation, 
+#' values are taken among \code{current}, \code{historic} and \code{all}. 
+#' Default is \code{current}.
+#' @param language vector of character strings indicating the language for the 
+#' text of legislation notes, values are taken among \code{en} (English), 
+#' \code{fr} (French) and \code{es} (Spanish). Default is \code{en}.
 #' @param simplify a logical. Should the output be simplified? In other words,
 #' should columns of data.table objects returned be unlisted when they are
 #' lists made of single elements?
@@ -29,8 +37,12 @@
 #' res2 <- spp_cites_legislation(taxon_id = '4521', type ='listings')
 #' }
 
-spp_cites_legislation <- function(taxon_id, type = c("listings", "quotas", "suspensions"), 
-    simplify = FALSE, token = NULL) {
+spp_cites_legislation <- function(taxon_id, 
+                                  type = c("listings", "quotas", "suspensions"), 
+                                  scope = c("current", "historic", "all"),
+                                  language = c("en", "fr", "es"),
+                                  simplify = FALSE, 
+                                  token = NULL) {
     # check token
     if (is.null(token)) 
         token <- rcites_getsecret()
@@ -38,8 +50,22 @@ spp_cites_legislation <- function(taxon_id, type = c("listings", "quotas", "susp
     stopifnot(all(type %in% c("listings", "quotas", "suspensions")))
     type <- unique(type)
     nmt <- c("listings", "quotas", "suspensions")
+    # set query_string
+    scope <- match.arg(scope)
+    if (scope == "current"){sc <- NULL}
+    if (scope == "historic"){sc <- "scope=historic"}
+    if (scope == "all"){sc <- "scope=all"}
+    language <- match.arg(language)
+    if (scope == "en"){la <- NULL}
+    if (scope == "fr"){la <- "language=fr"}
+    if (scope == "es"){la <- "language=es"}
+    query_string <- paste0(
+      if(is.null(sc) & is.null(la)){} else {"?"}, 
+      sc, 
+      if(!is.null(sc) & !is.null(la)){"&"} else {}, 
+      la)
     # 
-    q_url <- rcites_url(paste0("taxon_concepts/", taxon_id, "/cites_legislation.json"))
+    q_url <- rcites_url(paste0("taxon_concepts/", taxon_id, "/cites_legislation.json", query_string))
     res <- rcites_res(q_url, token)
     # output
     out <- lapply(res, function(x) "")
