+++ R/spp_taxonconcept.R
@@ -28,17 +28,17 @@
 #' @examples
 #' \donttest{
 #' res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
-#' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', appendix_only = TRUE)
+#' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', raw = TRUE)
 #' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
-#' res4 <- spp_taxonconcept(query_taxon = 'Amazilia')
+#' res4 <- spp_taxonconcept(query_taxon = 'Crotalus durissus')
 #' }
 
-spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL, raw = FALSE) {
+spp_taxonconcept <- function(query_taxon, token = NULL, raw = FALSE) {
     # token check
-    if (is.null(token)) 
+    if (is.null(token))
         token <- rcites_getsecret()
     query <- gsub(pattern = " ", replacement = "%20", x = query_taxon)
-    # 
+    #
     q_url <- rcites_url("taxon_concepts.json", "?name=", query)
     tmp <- rcites_res(q_url, token)
     pag <- rcites_numberpages(tmp$pagination)
@@ -47,23 +47,54 @@ spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL, r
         warning("Taxon not listed.")
         out <- NULL
     } else {
+        # auto pagination to be added here if (pag>1) rcites_autopagination()
+        # should be stored in tmp$taxon_concepts
         if (raw) {
             out <- tmp$taxon_concepts
             class(out) <- c("list", "spp_raw")
         } else {
-            sp_nm <- c("higher_taxa", "common_names", "synonyms", "cites_listings", 
-                "cites_listing", "accepted_names")
-            out <- do.call(rbind, lapply(tmp$taxon_concepts, function(x) unlist(x[!names(x) %in% 
-                sp_nm])))
-            # special cases return extra data frame sp_nm <- c('higher_taxa', 'common_names',
-            # 'synonyms', 'cites_listings', 'cites_listing', 'accepted_names') vc_nm <-
-            # unique(unlist(lapply(tmp2, names))) vc_nm <- vc_nm[! vc_nm %in% sp_nm] out <-
-            # lapply( lapply(tmp2, function(x) x[names(x) %in% vc_nm]), function(x) rbind(x,
-            # setdiff(vc_nm, names(x)) ))
-            class(out) <- c("tbl_df", "tbl", "data.frame", "spp_tax")
-            ## 
+            tmp <- tmp$taxon_concepts
+            ## special cases
+            sp_nm <- c("synonyms", "higher_taxa", "common_names",
+              "cites_listing",  "cites_listings", "accepted_names")
+            ## output
+            out <- list()
+            out$all_id <- data.frame(do.call(rbind, lapply(tmp,
+                function(x) unlist(x[!names(x) %in% sp_nm]))))
+            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at,
+                format = "%Y-%m-%dT%H:%M:%OS")
+            out$all_id$active <- as.logical(out$all_id$active)
+            ##
+            id <- which(out$all_id$active)
+            out$general <- out$all_id[out$all_id$active,]
+            out$general$cites_listing <- unlist(lapply(tmp[id], function(x) x$cites_listing))
+            class(out$all_id) <- c("tbl_df", "tbl", "data.frame")
+            class(out$general) <- c("tbl_df", "tbl", "data.frame")
+
[…]
