### 2018-08-25 27e4c91
minor format edit for better readability
