+++ R/spp_taxonconcept.R
@@ -3,15 +3,26 @@
 #' @description
 #' Retrieve the taxon concept of a specific taxon (scientific name).
 #'
-#' @param query_taxon a character string containing the query (e.g. species). Scientific taxa only (max 255 characters).
-#' @param appendix_only a logical. If \code{TRUE} then taxon identifier and the
-#' CITES appendix information are the only data returned (default is \code{FALSE}).
+#' @param query_taxon a character string containing the query (e.g. species).
+#' Scientific taxa only (max 255 characters).
+#' @param taxonomy filter taxon concepts by taxonomy, accepts either 'CITES' or
+#' 'CMS' as its value. Defaults set to 'CITES'.
+#' @param with_descendants a logical. Should the search by name be broadened to
+#' include higher taxa?
+#' @param language filter languages returned for common names. Value should be a
+#' vector of sting including one or more country codes. Default is set to `NULL`,
+# showing all available languages.
+#' @param per_page a integer that indicates how many objects are returned per
+#' page for paginated responses. Default set to 500 which is the maximum.
+#' @param seq_page a vector of integer that contains page numbers. Default is
+#' set to `NULL`, i.e. all pages are accessed.
+#' @param raw a logical. Should raw data be returned?
 #' @param token a character string containing the authentification token, see
 #' \url{https://api.speciesplus.net/documentation}. Default is set to
 #' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
 #' set directly in \code{Renviron}. Alternatively, \code{set_token()} can
 #' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
-#' @param raw a logical. Should raw data be returned?
+#' @param verbose a logical. Should extra information be reported on progress?
 #'
 #' @return
 #' If \code{appendix_only} is \code{TRUE}, then a data frame with a species'
@@ -30,54 +41,93 @@
 #' res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
 #' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', raw = TRUE)
 #' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
-#' res4 <- spp_taxonconcept(query_taxon = 'Crotalus durissus')
+#' res4 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = c(1,2))
+#' res5 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
 #' }
 
-spp_taxonconcept <- function(query_taxon, token = NULL, raw = FALSE) {
+spp_taxonconcept <- function(query_taxon, taxonomy = "CITES",
+  with_descendants = FALSE, language = NULL, per_page = 500, seq_page = NULL,
+  raw = FALSE, token = NULL, verbose = TRUE) {
     # token check
     if (is.null(token))
         token <- rcites_getsecret()
-    query <- gsub(pattern = " ", replacement = "%20", x = query_taxon)
-    #
-    q_url <- rcites_url("taxon_concepts.json", "?name=", query)
+    # taxonomy check
+    if (taxonomy == "CMS") {
+        taxo <- "CMS"
+    } else {
+        if (taxonomy != "CITES") {
+            warning("taxonomy only accepts either 'CITES' or 'CMS',
+        values set to 'CITES'")
+        }
+        taxo <- "CITES"
+    }
+    # request
+    if (is.null(seq_page)) {
+      f_page <- 1
+    } else {
[…]
