+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(print,spp_raw)
+S3method(print,spp_taxon)
 export(rcites_simplify)
 export(set_token)
 export(spp_cites_legislation)

[…]
