### 2018-08-27 1737d32
integrates Jonas comments see #33
