+++ R/spp_taxonconcept.R
@@ -41,7 +41,7 @@
 #' res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
 #' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', raw = TRUE)
 #' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
-#' res4 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = c(1,2))
+#' res4 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = 1:5, language = "EN")
 #' res5 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
 #' }
 
@@ -52,15 +52,7 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES",
     if (is.null(token))
         token <- rcites_getsecret()
     # taxonomy check
-    if (taxonomy == "CMS") {
-        taxo <- "CMS"
-    } else {
-        if (taxonomy != "CITES") {
-            warning("taxonomy only accepts either 'CITES' or 'CMS',
-        values set to 'CITES'")
-        }
-        taxo <- "CITES"
-    }
+    taxonomy <- match.arg(taxonomy, c("CITES", "CMS"))
     # request
     if (is.null(seq_page)) {
       f_page <- 1
@@ -68,8 +60,8 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES",
       seq_page <- sort(unique(as.integer(seq_page)))
       f_page <- unique(seq_page[1L])
     }
-    q_url <- rcites_request_taxonconcept(query_taxon, token, taxonomy = taxo,
-          with_descendants, f_page, per_page, language = NULL)
+    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy,
+          with_descendants, f_page, per_page, language)
     # results
     tmp <- rcites_res(q_url, token)
     # number of pages
@@ -103,7 +95,7 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES",
                 "cites_listings", "accepted_names")
             ## output
             out <- list()
-            out$all_id <- rcites_getallentries_taxonconcept(tmp2, sp_nm)
+            out$all_id <- rcites_taxonconcept_allentries(tmp2, sp_nm)
             out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at,
                 format = "%Y-%m-%dT%H:%M:%OS")
             out$all_id$active <- as.logical(out$all_id$active)
@@ -113,16 +105,19 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES",
 
 
             ## special cases NB add id
-            out$higher_taxa <- rcites_simplify_higher_taxa(tmp2[id], out$general$id)
-            out$synonyms <- rcites_simplify_special_cases(tmp2[id],
+            out$higher_taxa <- rcites_taxonconcept_higher_taxa(tmp2[id], out$general$id)
+            out$synonyms <- rcites_taxonconcept_special_cases(tmp2[id],
               name = "synonyms", out$general$id)
-            out$common_names <- rcites_simplify_special_cases(tmp2[id],
+            out$common_names <- rcites_taxonconcept_special_cases(tmp2[id],
               name = "common_names", out$general$id)
+            ##
+            out$accepted_names <- rcites_taxonconcept_special_cases(tmp2[!id],
+              name = "accepted_names", out$all_id$id[!id])
 
             ## Extra output if taxonomy is set to CITES
[…]
