+++ vignettes/rcites-vignette.Rmd
@@ -43,7 +43,7 @@ file (preferred);
 2. use `set_token()`, *i.e.* interactively set `SPECIESPLUS_TOKEN` up for the
 current R session (meaning a login will be required for the future R sessions);
 
-```{r}
+```{r, eval = FALSE}
 set_token("8QW6Qgh57sBG2k0gtt")
 ```
 
@@ -66,7 +66,7 @@ res1
 
 Note that if you chose options 3 regarding the connection:
 
-```{r}
+```{r, eval = FALSE}
 res1 <- spp_taxonconcept(query_taxon = "Loxodonta africana", token = "8QW6Qgh57sBG2k0gtt")
 ```
 
