### 2018-08-27 6671b14
eval=FALSE for the token to not cause an error anymore
