### 2018-08-27 20ba577
review scope and language
