+++ R/zzz.R
@@ -10,21 +10,21 @@
 
 ## Helper functions
 
-#
+# 
 rcites_baseurl <- function() "https://api.speciesplus.net/api/v1/"
 
-#
+# 
 rcites_url <- function(...) {
     paste0(rcites_baseurl(), ...)
 }
 
-#
+# 
 rcites_get <- function(q_url, token) {
     names(token) <- "X-Authentication-Token"
     httr::GET(q_url, httr::add_headers(token))
 }
 
-#
+# 
 rcites_res <- function(q_url, token) {
     con <- rcites_get(q_url, token)
     # check status
@@ -33,33 +33,43 @@ rcites_res <- function(q_url, token) {
     httr::content(con, "parsed")
 }
 
-#
+# 
 rcites_timestamp <- function(x) {
-  # ISO 8601 format
-  # https://stackoverflow.com/questions/29517896/current-time-in-iso-8601-format
-  tm <- as.POSIXlt(x, tz = "UTC")
-  strftime(tm , "%Y-%m-%dT%H:%M:%S")
+    # ISO 8601 format
+    # https://stackoverflow.com/questions/29517896/current-time-in-iso-8601-format
+    tm <- as.POSIXlt(x, tz = "UTC")
+    strftime(tm, "%Y-%m-%dT%H:%M:%S")
 }
 
-#
+# 
 rcites_lang <- function(x) {
-  out <- match.arg(x, c("en", "fr", "es"))
-  if (out == "en") out <- "" else out <- paste0("?language=", out)
-  out
+    out <- match.arg(x, c("en", "fr", "es"))
+    if (out == "en") 
+        out <- NULL else out <- paste0("language=", out)
+    out
+}
+
+# 
+rcites_scope <- function(x) {
+    out <- match.arg(x, c("current", "historic", "all"))
+    if (out == "current") 
+        out <- NULL else out <- paste0("scope=", out)
+    out
 }
 
 # auto pagination
-rcites_autopagination <- function(q_url, per_page, seq_page, tot_page,
+rcites_autopagination <- function(q_url, per_page, seq_page, tot_page, 
     token, verbose = TRUE) {
     out <- list()
-    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$",
+    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$", 
         replacement = "")
     # grepl('\\.json?', pat = '\\.json\\?$')
     for (i in seq_along(seq_page)) {
-        if (verbose)
-            cat("Retrieving info from page ", seq_page[i], "/", tot_page, "\r")
-        q_url_new <- paste0(q_url_0, "page=", seq_page[i], "&per_page=",
-        min(per_page, 500))
+        if (verbose) 
+            cat("Retrieving info from page ", seq_page[i], "/", tot_page, 
+                "\r")
+        q_url_new <- paste0(q_url_0, "page=", seq_page[i], "&per_page=", 
+            min(per_page, 500))
         out[[i]] <- rcites_res(q_url_new, token)
     }
     if (verbose) {
@@ -69,9 +79,9 @@ rcites_autopagination <- function(q_url, per_page, seq_page, tot_page,
     out
 }
 
-#
+# 
 rcites_print_df <- function(x, nrows = 10) {
-  print(x[seq_len(min(nrow(x), nrows)),])
+    print(x[seq_len(min(nrow(x), nrows)), ])
 }
 
 # See https://cran.r-project.org/web/packes/httr/vignettes/secrets.html
@@ -96,7 +106,7 @@ rcites_numberpages <- function(x) {
     x$total_entries%/%x$per_page + (x$total_entries%%x$per_page > 0)
 }
 
-# add 'author_year field
[…]
