+++ R/spp_taxonconcept.R
@@ -27,11 +27,7 @@
 #' @param verbose a logical. Should extra information be reported on progress?
 #'
 #' @return
-#' If \code{appendix_only} is \code{TRUE}, then a data frame with a species'
-#' taxon id and CITES appendix information is returned. Otherwise, a list of
-#' objects of class \code{data.table} with all CITES Species+ taxon_concept
-#' information is returned. Importantly enough, this functions returns the
-#' taxon concept identifier that is required by the \code{spp_*()} functions.
+#' To be added 
 #'
 #' @references
 #' \url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
@@ -47,11 +43,11 @@
 #' res5 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
 #' }
 
-spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE, 
-    language = NULL, updated_since = NULL, per_page = 500, seq_page = NULL, 
+spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE,
+    language = NULL, updated_since = NULL, per_page = 500, seq_page = NULL,
     raw = FALSE, token = NULL, verbose = TRUE) {
     # token check
-    if (is.null(token)) 
+    if (is.null(token))
         token <- rcites_getsecret()
     # taxonomy check
     taxonomy <- match.arg(taxonomy, c("CITES", "CMS"))
@@ -62,13 +58,13 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
         seq_page <- sort(unique(as.integer(seq_page)))
         f_page <- unique(seq_page[1L])
     }
-    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy, 
+    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy,
         with_descendants, f_page, per_page, updated_since, language)
     # results
     tmp <- rcites_res(q_url, token)
     # number of pages
     pag <- rcites_numberpages(tmp$pagination)
-    # 
+    #
     if (!pag) {
         warning("Taxon not listed.")
         out <- NULL
@@ -80,7 +76,7 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
                 seq_page <- seq_page[seq_page <= pag]
             }
             if (length(seq_page) > 1) {
-                res <- rcites_autopagination(q_url, per_page, seq_page[-1L], 
+                res <- rcites_autopagination(q_url, per_page, seq_page[-1L],
                   pag, token, verbose)
                 tmp2 <- c(tmp$taxon_concepts, do.call(c, lapply(res, function(x) x$taxon_concepts)))
             } else tmp2 <- tmp$taxon_concepts
@@ -92,42 +88,42 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
             class(out) <- c("list", "spp_raw")
         } else {
             ## special cases
-            sp_nm <- c("synonyms", "higher_taxa", "common_names", "cites_listing", 
+            sp_nm <- c("synonyms", "higher_taxa", "common_names", "cites_listing",
                 "cites_listings", "accepted_names")
             ## output
             out <- list()
             out$all_id <- rcites_taxonconcept_allentries(tmp2, sp_nm)
-            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at, 
+            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at,
                 format = "%Y-%m-%dT%H:%M:%OS")
             out$all_id$active <- as.logical(out$all_id$active)
             ## extract active only
             id <- which(out$all_id$active)
[…]
