+++ R/spp_distributions.R
@@ -8,6 +8,7 @@
 #' @param language vector of character strings indicating the language for the
 #' names of distributions, values are taken among \code{en} (English),
 #' \code{fr} (French) and \code{es} (Spanish). Default is \code{en}.
+#' @param raw a logical. Should raw data be returned?
 #' @param token a character string containing the authentification token, see
 #' \url{https://api.speciesplus.net/documentation}. Default is set to
 #' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
@@ -27,16 +28,15 @@
 #'  res2 <- spp_distributions(taxon_id = '4521', collapse_tags = ' + ')
 #' }
 
-spp_distributions <- function(taxon_id, language = "en", collapse_tags = NULL,
-    raw = FALSE, token = NULL) {
+spp_distributions <- function(taxon_id, language = "en", raw = FALSE, token = NULL) {
     # token check
-    if (is.null(token))
+    if (is.null(token)) 
         token <- rcites_getsecret()
     # set query_string
     tmp <- rcites_lang(language)
-    if (!is.null(tmp))
+    if (!is.null(tmp)) 
         tmp <- paste0("?", tmp)
-    q_url <- rcites_url("taxon_concepts/", taxon_id, "/distributions.json",
+    q_url <- rcites_url("taxon_concepts/", taxon_id, "/distributions.json", 
         tmp)
     # get results
     res <- rcites_res(q_url, token)
