### 2018-08-31 ecc7a0a
add a print method for object of class spp_cites_leg #34
