+++ R/spp_references.R
@@ -25,7 +25,7 @@
 
 spp_references <- function(taxon_id, raw = FALSE, token = NULL) {
     # token check
-    if (is.null(token)) 
+    if (is.null(token))
         token <- rcites_getsecret()
     ## create url
     q_url <- rcites_url("taxon_concepts/", taxon_id, "/references.json")
@@ -38,8 +38,8 @@ spp_references <- function(taxon_id, raw = FALSE, token = NULL) {
     } else {
         out <- list()
         out$references <- rcites_simplify_decisions(tmp)
-        class(out) <- c("spp_cites_ref")
+        class(out) <- c("spp_ref")
     }
-    # 
+    #
     out
 }
