+++ R/zzz.R
@@ -8,23 +8,21 @@
 #' @keywords internal
 "_PACKAGE"
 
-## Helper functions
+# Helper functions
 
-# 
+## General helpers
+
 rcites_baseurl <- function() "https://api.speciesplus.net/api/v1/"
 
-# 
 rcites_url <- function(...) {
     paste0(rcites_baseurl(), ...)
 }
 
-# 
 rcites_get <- function(q_url, token) {
     names(token) <- "X-Authentication-Token"
     httr::GET(q_url, httr::add_headers(token))
 }
 
-# 
 rcites_res <- function(q_url, token) {
     con <- rcites_get(q_url, token)
     # check status
@@ -33,7 +31,6 @@ rcites_res <- function(q_url, token) {
     httr::content(con, "parsed")
 }
 
-# 
 rcites_timestamp <- function(x) {
     # ISO 8601 format
     # https://stackoverflow.com/questions/29517896/current-time-in-iso-8601-format
@@ -41,48 +38,22 @@ rcites_timestamp <- function(x) {
     strftime(tm, "%Y-%m-%dT%H:%M:%S")
 }
 
-# 
 rcites_lang <- function(x) {
     out <- match.arg(x, c("en", "fr", "es"))
-    if (out == "en") 
+    if (out == "en")
         out <- NULL else out <- paste0("language=", out)
     out
 }
 
-# 
 rcites_scope <- function(x) {
     out <- match.arg(x, c("current", "historic", "all"))
-    if (out == "current") 
+    if (out == "current")
         out <- NULL else out <- paste0("scope=", out)
     out
 }
 
-# auto pagination
-rcites_autopagination <- function(q_url, per_page, seq_page, tot_page, 
-    token, verbose = TRUE) {
-    out <- list()
-    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$", 
-        replacement = "")
-    # grepl('\\.json?', pat = '\\.json\\?$')
-    for (i in seq_along(seq_page)) {
-        if (verbose) 
-            cat("Retrieving info from page ", seq_page[i], "/", tot_page, 
-                "\r")
-        q_url_new <- paste0(q_url_0, "page=", seq_page[i], "&per_page=", 
-            min(per_page, 500))
-        out[[i]] <- rcites_res(q_url_new, token)
-    }
-    if (verbose) {
-        cat("\n")
-        cat("Done!\n")
-    }
-    out
-}
 
-# 
-rcites_print_df <- function(x, nrows = 10) {
-    print(x[seq_len(min(nrow(x), nrows)), ])
-}
+## Secret helpers
 
 # See https://cran.r-project.org/web/packes/httr/vignettes/secrets.html
 rcites_getsecret <- function() {
@@ -101,10 +72,39 @@ rcites_getsecret <- function() {
 # remove secret
 rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 
+
+## Pagination helpers
+
+# auto pagination
+rcites_autopagination <- function(q_url, per_page, seq_page, tot_page,
+    token, verbose = TRUE) {
+    out <- list()
+    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$",
+        replacement = "")
+    # grepl('\\.json?', pat = '\\.json\\?$')
+    for (i in seq_along(seq_page)) {
+        if (verbose)
+            cat("Retrieving info from page ", seq_page[i], "/", tot_page,
+                "\r")
+        q_url_new <- paste0(q_url_0, "page=", seq_page[i], "&per_page=",
+            min(per_page, 500))
[…]
