+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,spp_cites_leg)
 S3method(print,spp_eu_leg)
 S3method(print,spp_raw)
 S3method(print,spp_taxon)

[…]
