### 2018-08-31 349ca3e
add print method for spp_refs  #34
