+++ R/spp_references.R
@@ -8,7 +8,7 @@
 #' @param token a character string containing the authentification token, see
 #' \url{https://api.speciesplus.net/documentation}. Default is set to
 #' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
-#' set directly in \code{Renviron}. Alternatively \code{set_token()} can
+#' set directly in \code{Renviron}. Alternatively, \code{set_token()} can
 #' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
 #'
 #' @return A list of objects, one per type requested.
@@ -21,6 +21,7 @@
 #' @examples
 #' \donttest{
 #' res1 <- spp_references(taxon_id = '4521')
+#' res2 <- spp_references(taxon_id = '4521', raw = TRUE)
 #' }
 
 spp_references <- function(taxon_id, raw = FALSE, token = NULL) {
