+++ man/spp_distributions.Rd
@@ -33,7 +33,7 @@ the taxon identifier is known.
 \examples{
 \donttest{
  res1 <- spp_distributions(taxon_id = '4521')
- res2 <- spp_distributions(taxon_id = '4521', collapse_tags = ' + ')
+ res2 <- spp_distributions(taxon_id = '4521', raw = TRUE)
 }
 }
 \references{
