+++ DESCRIPTION
@@ -21,5 +21,5 @@ Imports:
   methods
 Encoding: UTF-8
 RoxygenNote: 6.1.0
-Suggests: knitr, rmarkdown, testthat, rworldmap
+Suggests: knitr, testthat, tibble, rmarkdown, rworldmap
 VignetteBuilder: knitr
