### 2018-08-31 ae6e1e8
add print method for spp_distr  #34
