+++ R/zzz.R
@@ -40,14 +40,14 @@ rcites_timestamp <- function(x) {
 
 rcites_lang <- function(x) {
     out <- match.arg(x, c("en", "fr", "es"))
-    if (out == "en")
+    if (out == "en") 
         out <- NULL else out <- paste0("language=", out)
     out
 }
 
 rcites_scope <- function(x) {
     out <- match.arg(x, c("current", "historic", "all"))
-    if (out == "current")
+    if (out == "current") 
         out <- NULL else out <- paste0("scope=", out)
     out
 }
@@ -76,17 +76,17 @@ rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 ## Pagination helpers
 
 # auto pagination
-rcites_autopagination <- function(q_url, per_page, seq_page, tot_page,
+rcites_autopagination <- function(q_url, per_page, seq_page, tot_page, 
     token, verbose = TRUE) {
     out <- list()
-    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$",
+    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$", 
         replacement = "")
     # grepl('\\.json?', pat = '\\.json\\?$')
     for (i in seq_along(seq_page)) {
-        if (verbose)
-            cat("Retrieving info from page ", seq_page[i], "/", tot_page,
+        if (verbose) 
+            cat("Retrieving info from page ", seq_page[i], "/", tot_page, 
                 "\r")
-        q_url_new <- paste0(q_url_0, "page=", seq_page[i], "&per_page=",
+        q_url_new <- paste0(q_url_0, "page=", seq_page[i], "&per_page=", 
             min(per_page, 500))
         out[[i]] <- rcites_res(q_url_new, token)
     }
@@ -114,12 +114,12 @@ rcites_addauthor <- function(x) {
     x
 }
 
-#
+# 
 rcites_simplify_listings <- function(x) {
     # these fields may or may not be included, so I removed them
     vc_nm <- c("party", "hash_annotation", "annotation")
     tmp <- lapply(x, function(y) data.frame(do.call(cbind, y)))
-    out <- do.call(rbind, lapply(tmp, function(y) y[, which(!names(y) %in%
+    out <- do.call(rbind, lapply(tmp, function(y) y[, which(!names(y) %in% 
         vc_nm)]))
     rownames(out) <- NULL
     class(out) <- c("tbl_df", "tbl", "data.frame")
@@ -135,7 +135,7 @@ r_cites_null_to_na <- function(x) {
     }
 }
 
-#
+# 
 rcites_simplify_decisions <- function(x) {
     # these fields may or may not be included so I removed them
     tmp0 <- lapply(lapply(x, r_cites_null_to_na), unlist)
@@ -145,22 +145,22 @@ rcites_simplify_decisions <- function(x) {
     out
 }
 
-#
+# 
 rcites_simplify_distributions <- function(x) {
     # these fields may or may not be included so I removed them
     tmp <- lapply(x, r_cites_null_to_na)
     out <- list()
-    out$distributions <- data.frame(do.call(rbind, lapply(tmp, function(y) data.frame(rbind(unlist(y[!names(y) %in%
+    out$distributions <- data.frame(do.call(rbind, lapply(tmp, function(y) data.frame(rbind(unlist(y[!names(y) %in% 
         c("tags", "references")]))))))
-    out$distributions$tags <- unlist(lapply(tmp, function(y) paste(y$tags,
[…]
