+++ R/spp_references.R
@@ -26,7 +26,7 @@
 
 spp_references <- function(taxon_id, raw = FALSE, token = NULL) {
     # token check
-    if (is.null(token))
+    if (is.null(token)) 
         token <- rcites_getsecret()
     ## create url
     q_url <- rcites_url("taxon_concepts/", taxon_id, "/references.json")
@@ -41,6 +41,6 @@ spp_references <- function(taxon_id, raw = FALSE, token = NULL) {
         out$references <- rcites_simplify_decisions(tmp)
         class(out) <- c("spp_refs")
     }
-    #
+    # 
     out
 }
