### 2018-08-31 bfa43ee
unit testing for helper functions #34
