### 2018-09-07 #244 @mcsiple
Hi @JonasGeschke et al. - I am having the same compiling issues as @noamross with the `.Rmd` files for vignettes, and it would be helpful for using and reviewing if the outputs were visible (either by pasting in blocks or putting a pdf in the vignettes folder, if API key issues happen often). I did a temp fix by adding my own API key to the `.Rmd` file but had to add the token to each line with an `rcites` function. If other reviewers are assigned (and if it is in accordance with rOpenSci reviewing practices) it might be helpful to make a compiled pdf or html available.
