### 2018-09-07 #244 @JonasGeschke
Hi @mcsiple thank you for your post. Yes, making the vignette more comprehensive and with outputs visible is on our list. If you have any further reviewing comments, we are happy to receive those. Thank you!
