### 2018-09-07 #244 @mcsiple
The rest of my comment are coming soon - I just wanted to post that one in case other reviewers are assigned and need the compiled vignettes in the meantime!
