### 2018-09-10 #244 @mcsiple
-  Currently the vignette is included as a file that doesn't compile locally because of API issues (as mentioned by @noamross).

>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I found this package lovely, easy to use, and well-documented. I think the authors have done a good job of wrangling some complex data types from the API and am sure this package will be widely used. Most of the suggestions I have are about interfacing with the API, as well as a few about usability.

Once compiled, the vignette is quite concise and nicely written (There are a few typos in the vignette but none that affect functionality or clarity). Some parts of the vignette could be fleshed out a little more: for example, the "Legislation information" section does not have text to accompany the code. When I compiled the vignette, the `spp_cites_legislation()` piece here returns an "invalid factor level" warning, and some accompanying text would help me as a user figure out what I am supposed to see here, and whether it is an issue with the vignette or with the code.

The README is also good.

As I mentioned in my earlier comment, the API token needs to be set in the environment before the `.Rmd` file will compile (i.e., even if a valid token is entered at the beginning, `set_token()` has to be used before the vignette will knit).

## Functions

Setting the token: `set_token()` example should note that token is entered as a character string if it is within the function, and without quotes (not a character object) when it's entered after the prompt.
E.g.,
`set_token("VrXZVlHjhEsVjerBfqsV5Qtt")`
or `set_token()`
`Enter your token: VrXZVlHjhEsVjerBfqsV5Qtt`
If the token is entered as a character object with the prompt, the token will not set properly and the dependent functions don't work.

There are a lot of `rcites_` functions within the package functions (e.g., `rcites_getsecret()` which is used in `spp_taxonconcept()`) which don't work on their own. Since the wrapper functions around them work, I think it's fine; I was just confused as to what they are. I'm not sure what the rOpenSci guidelines are for these functions.

## Icing to add: improving processing time
One of the greatest challenges I have with any package that uses an API is the processing time to get bulk entries.

For example, I usually have a dataframe for which I need to access and merge data I've gathered through something like `rcites`. While I found that `rcites` returns objects that are easy to extract, I think it would be particularly wonderful if there were a few ways to access that data without having the re-look them up with the API every time. For example, if the same list of species is looked up multiple times, could those data be stored locally for faster access if the user wants to look up something else about them? I think @noamross also mentions this in his review. This would also be helpful so authors don't overload the API. I don't know if the CITES API limits how many data requests each user can make at a time, but this added functionality would help avoid that issue.

Finally, I also think the authors should make tibble output an option for all the users who will be using `dplyr` with `rcites` (I would!).

## Mapping

I love that code for making distribution maps is included in the vignette, but had some issues with the example. The mapping example in the vignette doesn't work- the following error occurs when running `as.data.frame(spp_distributions("4521"))`:
`"cannot coerce class ‘"spp_distr"’ to a data.frame"`

I know that this is not the main point of the package, but it's cool that there is an easy way to make these maps with `rcites` output, so I would love to see the outputs from this example.

rOpenSci guidelines:
- [X] Does the code comply with general principles in the Mozilla reviewing guide?
- [X] Does the package comply with the rOpenSci packaging guide?
