### 2018-09-10 #244 @KevCaz
Many thanks @mcsiple for all your positive comments, much appreciated :smile: .
We are working on a new version that integrates all the reviewers comments.
