### 2018-09-10 #244 @JonasGeschke
Yes indeed, thank you so much @mcsiple! I will add your comments to our list of to-dos and we will continue to work on them. we hope to report back to you asap.
