### 2018-09-10 da9c0de
set_token explanation, associated with ROpenSci review
