+++ vignettes/get_started.Rmd
@@ -43,8 +43,11 @@ requests. They are three different ways to set the token in `rcites`:
 [`.Renviron`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/Startup.html)
 file (preferred);
 
-2. use `set_token()`, *i.e.* interactively set `SPECIESPLUS_TOKEN` up for the
-current R session (meaning a login will be required for the future R sessions);
+2. use `set_token()` to set up the token as character string (with quotes). If not 
+entered in the function, the token can be passed without quotes (not as character 
+string) after a prompt. This way, the token `SPECIESPLUS_TOKEN` is interactively 
+set up for the current R session (meaning a login will be required for the future 
+R sessions);
 
 ```{r, eval = FALSE}
 set_token("8QW6Qgh57sBG2k0gtt")
