+++ vignettes/rcites-vignette.Rmd
@@ -1,162 +1,0 @@
----
-title: "Get started with rcites"
-author: "rcites team"
-date: 07-18-2018
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Get started with rcites}
-  %\VignetteEngine{knitr::rmarkdown}
----
-
-
-```{r setup, echo = FALSE}
-NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
-knitr::opts_chunk$set(
-  purl = NOT_CRAN,
-  eval = NOT_CRAN,
-  fig.align = "center",
-  comment = "#> "
-)
-```
-
-
-<!-- https://cran.r-project.org/web/packages/httr/vignettes/secrets.html -->
-
-
-## Set up a connection to the species plus database
-
-To set up a connection to the CITES Speciesplus database, an authentication token
-is required. To obtain such a token, the user must sign up on the
-[Speciesplus API website](http://api.speciesplus.net/users/sign_up).
-Below, we assume that the user already possesses a token. For illustrative purposes
-we will use the generic token value `8QW6Qgh57sBG2k0gtt` just as on the API
-documentation, but each user should obtain his own token to run the code below (see
-https://api.speciesplus.net/documentation for more details).
-
-A token is actually mandatory as it should be passed to the header of all URL
-requests. They are three different ways to use such token in `rcites`:
-
-1. set an environment variable `SPECIESPLUS_TOKEN` in
-[`.Renviron`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/Startup.html)
-file (preferred);
-
-2. use `set_token()`, *i.e.* interactively set `SPECIESPLUS_TOKEN` up for the
-current R session (meaning a login will be required for the future R sessions);
-
-```{r, eval = FALSE}
-set_token("8QW6Qgh57sBG2k0gtt")
-```
-
-3. use the `token` argument of the functions, *i.e.* the token is passed manually to every function calls.
-
-
-## Retrieve a taxon concept identifier
-
-To efficiently query information from the CITES Speciesplus database, the user
-first needs to retrieve the unique identifier of taxon concept considered (https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html). To do
-so, the user should first call `spp_taxonconcept()` and provide the scientific
-name of the taxon. Let us start by requesting the identifier of the
-African bush elephant, *i.e.* *Loxodonta africana*.
-
-```{r}
-library(rcites)
-res1 <- spp_taxonconcept(query_taxon = "Loxodonta africana")
-res1
-```
-
-Note that if you chose options 3 regarding the connection:
-
-```{r, eval = FALSE}
-res1 <- spp_taxonconcept(query_taxon = "Loxodonta africana", token = "8QW6Qgh57sBG2k0gtt")
-```
-
-`res1` is a list of data.table object with different information pertaining
-to the naming of the taxon. If you however only need many details, setting
-`appendix_only` to `TRUE` will return the first data.table, `all`, for instance:
-
-```{r qtxla2}
-res2 <- spp_taxonconcept(query_taxon = "Loxodonta africana")
-res2
-```
-
-For other taxon, there might be more than one identifier, in such a case the user
-should have a look at the `active` column.
-
-```{r qtxav}
-res3 <- spp_taxonconcept(query = "Amazilia versicolor")
-res3
-```
-
-Also, if the taxon is not listed a warning message should come up:
-
-```{r}
-res <- spp_taxonconcept(query = "Homo Sapiens")
-```
-
-
[…]
