### 2018-09-10 f9e3d7f
function prefix explanation, associated with ROpenSci review
