### 2018-09-10 1b5afdc
minor rewording
