+++ R/set_token.R
@@ -12,12 +12,15 @@
 #'
 #' @examples
 #' \donttest{
+#'  # NB the token below is not working
+#'  set_token("8QW6Qgh57sBG2k0gtt")
+#'  # interactively
 #'  set_token()
 #' }
 
 set_token <- function(token = NULL) {
-    if (is.null(token)) 
-        token <- readline("Enter your token: ")
+    if (is.null(token))
+        token <- readline("Enter your token without quotes: ")
     if (identical(token, "")) {
         message("no token provided")
     } else {
