+++ DESCRIPTION
@@ -8,7 +8,8 @@ Authors@R: c(person("Jonas", "Geschke", role = c("aut"), email = "jonas.e.geschk
   person("Jonathan", "Goldenberg", role = c("ctb")),
   person("Marie-Bé", "Leduc", role = c("ctb")),
   person("Yasmine", "Verzelen", role = c("ctb")),
-  person("Noam", "Ross", role = c("rev")))
+  person("Noam", "Ross", role = c("rev")),
+  person("Margaret", "Siple", role = c("rev")))
 Description: A programmatic interface to the Species+ <https://speciesplus.net/> database via the Species+/CITES Checklist API <https://api.speciesplus.net/>.
 URL: https://ibartomeus.github.io/rcites/, https://github.com/ibartomeus/rcites
 BugReports: https://github.com/ibartomeus/rcites/issues
