### 2018-09-16 548d147
few modification in the print methods
