+++ R/spp_taxonconcept.R
@@ -12,7 +12,7 @@
 #' @param language filter languages returned for common names. Value should be a
 #' vector of sting including one or more country codes. Default is set to `NULL`,
 #' showing all available languages.
-#' @param updated_since a timestamp. Only objects updated after (and including)
+#' @param updated_since a timestamp. Only entries updated after (and including)
 #' this timestamp will be pulled.
 #' @param per_page a integer that indicates how many objects are returned per
 #' page for paginated responses. Default set to 500 which is the maximum.
@@ -44,20 +44,19 @@
 #' @examples
 #' \donttest{
 #' res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
-#' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', raw = TRUE)
-#' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
-#' res4 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = 1:5, language = 'EN')
-#' res5 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
+#' res2 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor', raw = TRUE)
+#' res3 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = 1:5, language = 'EN')
+#' res4 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
 #' }
 
-spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE, 
-    language = NULL, updated_since = NULL, per_page = 500, seq_page = NULL, 
+spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE,
+    language = NULL, updated_since = NULL, per_page = 500, seq_page = NULL,
     raw = FALSE, token = NULL, verbose = TRUE) {
-    # token check
-    if (is.null(token)) 
-        token <- rcites_getsecret()
     # taxonomy check
     taxonomy <- match.arg(taxonomy, c("CITES", "CMS"))
+    # token check
+    if (is.null(token))
+        token <- rcites_getsecret()
     # request
     if (is.null(seq_page)) {
         f_page <- 1
@@ -65,13 +64,13 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
         seq_page <- sort(unique(as.integer(seq_page)))
         f_page <- unique(seq_page[1L])
     }
-    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy, 
+    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy,
         with_descendants, f_page, per_page, updated_since, language)
     # results
     tmp <- rcites_res(q_url, token)
     # number of pages
     pag <- rcites_numberpages(tmp$pagination)
-    # 
+    #
     if (!pag) {
         warning("Taxon not listed.")
         return(NULL)
@@ -83,7 +82,7 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
                 seq_page <- seq_page[seq_page <= pag]
             }
             if (length(seq_page) > 1) {
-                res <- rcites_autopagination(q_url, per_page, seq_page[-1L], 
+                res <- rcites_autopagination(q_url, per_page, seq_page[-1L],
                   pag, token, verbose)
                 tmp2 <- c(tmp$taxon_concepts, do.call(c, lapply(res, function(x) x$taxon_concepts)))
             } else tmp2 <- tmp$taxon_concepts
[…]
