+++ tests/testthat/test-taxonconcept.R
@@ -1,43 +1,50 @@
 context("taxon concept")
 
+## General helpers
+test_that("General helpers", {
+  expect_error(spp_taxonconcept(query_taxon = tx_nm, taxonomy = "WRONG"))
+})
+
 skip_on_cran()
 skip_if_no_auth()
 
-# ut_pause()
-# res1 <- spp_taxonconcept(query_taxon = tx_nm)
+res1 <- spp_taxonconcept(query_taxon = tx_nm)
+nm1 <- c("all_id", "general", "higher_taxa", "synonyms", "common_names",
+  "accepted_names", "cites_listings")
+nm2 <- c("id", "full_name", "author_year", "rank", "name_status", "updated_at",
+  "active", "cites_listing")
+ut_pause()
+
+res2 <- spp_taxonconcept(query_taxon = tx_nm, raw = TRUE)
+ut_pause()
+
+res3 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = 1:2,
+  language = 'EN', verbose = FALSE)
+ut_pause()
+
+res4 <- spp_taxonconcept(query_taxon = '', seq_page = c(44), with_descendants = FALSE)
+ut_pause()
+
+res5 <- spp_taxonconcept(query_taxon = '', seq_page = 1, updated_since = "2016-01-01")
+ut_pause()
 
 
+test_that("Expected classes", {
+  expect_equal(class(res1), "spp_taxon")
+  expect_equal(length(res1), 7)
+  expect_true(all(names(res1) == nm1))
+  expect_true(class(res1[1L]) == "list")
+  expect_true(all(names(res1[[2L]]) == nm2))
+  expect_true(all(unlist(lapply(res1, function(x) all(class(x) == cl_df)))))
+  expect_true(all(class(res2) == cl_raw))
+})
 
-# ut_pause()
-# res2 <- spp_taxonconcept(query_taxon = tx_nm, appendix_only = TRUE)
-# nm1 <- c("id", "full_name", "author_year", "rank", "name_status", "updated_at", "active", "cites_listing")
-# nm2 <- c("all", "synonyms", "common_names", "higher_taxa")
-# ut_pause()
-# res_null <- suppressWarnings(spp_taxonconcept(query_taxon = "Homo sapiens"))
-# ut_pause()
-# res3 <- spp_taxonconcept(query_taxon = "Amazilia versicolor")
-#
-#
-# test_that("expected classes", {
-#   expect_equal(class(res1), "list")
-#   expect_true(all(class(res1$all) == cl_dt))
-#   expect_true(all(class(res1$synonyms) == cl_dt))
-#   expect_true(all(class(res1$higher_taxa) == cl_dt))
-#   expect_true(all(class(res1$common_names) == cl_dt))
-#   expect_true(all(class(res2) == cl_dt))
-#   expect_true(is.null(res_null))
-#   ut_pause()
-#   expect_warning(spp_taxonconcept(query_taxon = "Homo sapiens"), "Taxon not listed.", fixed = TRUE)
-# })
-#
-# test_that("expected output names", {
-#   expect_true(all(names(res1) == nm2))
-#   expect_true(all(names(res1$all) == nm1))
-#   expect_true(all(names(res2) == nm1))
-# })
-#
-# test_that("expected values", {
-#   expect_equal(res2$id, tx_id)
-#   expect_equal(res2$full_name, tx_nm)
-#   expect_true(all(dim(res3$all) == c(2, 8)))
-# })
+
+test_that("Expected behaviour", {
+  expect_equal(nrow(res3[[1L]]), 1000)
+  expect_true(all(res3$common_names$language == "EN"))
[…]
