### 2018-09-16 b410ca6
add test for distribution
