### 2018-09-16 4fabd02
better document spp_taxonconcept() outputs
