+++ tests/testthat/test-taxonconcept.R
@@ -9,8 +9,8 @@ skip_on_cran()
 skip_if_no_auth()
 
 res1 <- spp_taxonconcept(query_taxon = tx_nm)
-nm1 <- c("all_id", "general", "higher_taxa", "synonyms", "common_names",
-  "accepted_names", "cites_listings")
+nm1 <- c("all_id", "general", "higher_taxa", "accepted_names", "common_names",
+"synonyms", "cites_listings")
 nm2 <- c("id", "full_name", "author_year", "rank", "name_status", "updated_at",
   "active", "cites_listing")
 ut_pause()
