+++ tests/testthat/test-wrongtoken.R
@@ -1,9 +1,9 @@
-context("wong token")
+context("Wrong token")
 
-token <- "wrong"
-msg <- "Unauthorized (HTTP 401)"
+# token <- "wrong"
+# msg <- "Unauthorized (HTTP 401)"
 # q_url <- rcites_url(paste0("taxon_concepts/", tx_id, "/distributions.json"))
-
+#
 # test_that("expected url", {
 #   expect_equal(q_url, "https://api.speciesplus.net/api/v1/taxon_concepts/4521/distributions.json")
 # })
