### 2018-09-16 4d35348
add rcites_checkid()
