+++ tests/testthat/test-legislations.R
@@ -3,17 +3,17 @@ context("Legislations")
 skip_on_cran()
 skip_if_no_auth()
 #
-# res1 <- spp_cites_legislation(taxon_id = tx_nm)
-# res2 <- spp_cites_legislation(taxon_id = tx_nm, raw = TRUE)
+# res1 <- spp_cites_legislation(taxon_id = tx_id)
+# res2 <- spp_cites_legislation(taxon_id = tx_id, raw = TRUE)
 # ut_pause()
-# res3 <- spp_cites_legislation(taxon_id = tx_nm, scope = 'all')
-# res4 <- spp_cites_legislation(taxon_id = tx_nm, language = 'fr')
+# res3 <- spp_cites_legislation(taxon_id = tx_id, scope = 'all')
+# res4 <- spp_cites_legislation(taxon_id = tx_id, language = 'fr')
 # ut_pause()
 #
-# res5 <- spp_eu_legislation(taxon_id = tx_nm)
-# res6 <- spp_eu_legislation(taxon_id = tx_nm, raw = TRUE)
-# res7 <- spp_eu_legislation(taxon_id = tx_nm, scope = 'all')
-# res8 <- spp_eu_legislation(taxon_id = tx_nm, language = 'fr')
+# res5 <- spp_eu_legislation(taxon_id = tx_id)
+# res6 <- spp_eu_legislation(taxon_id = tx_id, raw = TRUE)
+# res7 <- spp_eu_legislation(taxon_id = tx_id, scope = 'all')
+# res8 <- spp_eu_legislation(taxon_id = tx_id, language = 'fr')
 # ut_pause()
 
 # #
