+++ R/spp_taxonconcept.R
@@ -36,7 +36,7 @@
 #' 4. `accepted_names`: list of accepted names (only for synonyms),
 #' 5. `common_names`: list of common names (only for accepted names),
 #' 6. `synonyms`: list of synonyms (only for accepted names),
-#' 7. `cites_listing`: list of current CITES listings with annotations (missing if `taxonomy == "CMS"`).
+#' 7. `cites_listing`: list of current CITES listings with annotations (missing if `taxonomy == 'CMS'`).
 #'
 #' @references
 #' \url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
@@ -51,13 +51,13 @@
 #' res4 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
 #' }
 
-spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE,
-    language = NULL, updated_since = NULL, per_page = 500, seq_page = NULL,
+spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE, 
+    language = NULL, updated_since = NULL, per_page = 500, seq_page = NULL, 
     raw = FALSE, token = NULL, verbose = TRUE) {
     # taxonomy check
     taxonomy <- match.arg(taxonomy, c("CITES", "CMS"))
     # token check
-    if (is.null(token))
+    if (is.null(token)) 
         token <- rcites_getsecret()
     # request
     if (is.null(seq_page)) {
@@ -66,13 +66,13 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
         seq_page <- sort(unique(as.integer(seq_page)))
         f_page <- unique(seq_page[1L])
     }
-    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy,
+    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy, 
         with_descendants, f_page, per_page, updated_since, language)
     # results
     tmp <- rcites_res(q_url, token)
     # number of pages
     pag <- rcites_numberpages(tmp$pagination)
-    #
+    # 
     if (!pag) {
         warning("Taxon not listed.")
         return(NULL)
@@ -84,7 +84,7 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
                 seq_page <- seq_page[seq_page <= pag]
             }
             if (length(seq_page) > 1) {
-                res <- rcites_autopagination(q_url, per_page, seq_page[-1L],
+                res <- rcites_autopagination(q_url, per_page, seq_page[-1L], 
                   pag, token, verbose)
                 tmp2 <- c(tmp$taxon_concepts, do.call(c, lapply(res, function(x) x$taxon_concepts)))
             } else tmp2 <- tmp$taxon_concepts
@@ -95,42 +95,42 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
             return(tmp2)
         } else {
             ## special cases
-            sp_nm <- c("synonyms", "higher_taxa", "common_names", "cites_listing",
+            sp_nm <- c("synonyms", "higher_taxa", "common_names", "cites_listing", 
                 "cites_listings", "accepted_names")
             ## output
             out <- list()
             out$all_id <- rcites_taxonconcept_allentries(tmp2, sp_nm)
-            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at,
+            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at, 
                 format = "%Y-%m-%dT%H:%M:%OS")
             out$all_id$active <- as.logical(out$all_id$active)
             ## extract active only
             id <- which(out$all_id$active)
             out$general <- out$all_id[out$all_id$active, ]
-
[…]
