### 2018-09-16 e2b6385
use Markdown to document the functions
