+++ R/spp_distributions.R
@@ -4,16 +4,16 @@
 #' the taxon identifier is known.
 #'
 #' @param taxon_id character string containing a species' taxon concept identifier
-#' (see \code{\link[rcites]{spp_taxonconcept}}).
+#' (see [spp_taxonconcept()]).
 #' @param language vector of character strings indicating the language for the
-#' names of distributions, values are taken among \code{en} (English),
-#' \code{fr} (French) and \code{es} (Spanish). Default is \code{en}.
+#' names of distributions, values are taken among `en` (English),
+#' `fr` (French) and `es` (Spanish). Default is `en`.
 #' @param raw a logical. Should raw data be returned?
 #' @param token a character string containing the authentification token, see
-#' \url{https://api.speciesplus.net/documentation}. Default is set to
-#' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
-#' set directly in \code{Renviron}. Alternatively \code{set_token()} can
-#' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
+#' <https://api.speciesplus.net/documentation>. Default is set to
+#' `NULL` and requires the environment variable `SPECIESPLUS_TOKEN` to be
+#' set directly in `Renviron`. Alternatively, \code{set_token()} can
+#' be used to set `SPECIESPLUS_TOKEN` for the current session.
 #'
 #' @return A data frame with all distribution information.
 #'
@@ -31,15 +31,15 @@
 
 spp_distributions <- function(taxon_id, language = "en", raw = FALSE, token = NULL) {
     # token check
-    if (is.null(token)) 
+    if (is.null(token))
         token <- rcites_getsecret()
     # id check
     rcites_checkid(taxon_id)
     # set query_string
     tmp <- rcites_lang(language)
-    if (!is.null(tmp)) 
+    if (!is.null(tmp))
         tmp <- paste0("?", tmp)
-    q_url <- rcites_url("taxon_concepts/", taxon_id, "/distributions.json", 
+    q_url <- rcites_url("taxon_concepts/", taxon_id, "/distributions.json",
         tmp)
     # get results
     res <- rcites_res(q_url, token)
