### 2018-09-16 a32e538
better document functions output
