+++ .travis.yml
@@ -9,11 +9,13 @@ matrix:
     - os: linux
       r: oldrel
       env: NOT_CRAN=false
+      r_check_args: "--no-vignettes"
     - os: linux
       r: release
       env: R_CODECOV=true
     - os: linux
       r: devel
+      r_check_args: "--no-vignettes"
 
 
 notifications:
