### 2018-09-16 4a726da
few changes in tests
