### 2018-09-17 b5965f9
Newtests (#35)
* few modification in the print methods

* test spp_taxonconcept

* add test for distribution

* better document spp_taxonconcept() outputs

* add rcites_checkid()

* add test-wrongtoken back

* test references

* test legislations :shield:

* use Markdown to document the functions

* better document functions output

* rebuild the website

* non vignette in oldrel and devel

* few changes in tests

* few changes in .travis.yml

* rebuild website :book:
