+++ R/spp_taxonconcept.R
@@ -6,28 +6,37 @@
 #' @param query_taxon a character string containing the query (e.g. species).
 #' Scientific taxa only (max 255 characters).
 #' @param taxonomy filter taxon concepts by taxonomy, accepts either 'CITES' or
-#' 'CMS' as its value. Defaults set to 'CITES'.
+#' 'CMS' as its value. Default sets to 'CITES'.
 #' @param with_descendants a logical. Should the search by name be broadened to
 #' include higher taxa?
 #' @param language filter languages returned for common names. Value should be a
 #' vector of sting including one or more country codes. Default is set to `NULL`,
 #' showing all available languages.
-#' @param updated_since a timestamp. Only objects updated after (and including)
+#' @param updated_since a timestamp. Only entries updated after (and including)
 #' this timestamp will be pulled.
 #' @param per_page a integer that indicates how many objects are returned per
 #' page for paginated responses. Default set to 500 which is the maximum.
-#' @param seq_page a vector of integer that contains page numbers. Default is
+#' @param pages a vector of integer that contains page numbers. Default is
 #' set to `NULL`, i.e. all pages are accessed.
 #' @param raw a logical. Should raw data be returned?
 #' @param token a character string containing the authentification token, see
 #' \url{https://api.speciesplus.net/documentation}. Default is set to
-#' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
-#' set directly in \code{Renviron}. Alternatively, \code{set_token()} can
-#' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
+#' `NULL` and requires the environment variable `SPECIESPLUS_TOKEN` to be
+#' set directly in `Renviron`. Alternatively, `set_token()` can
+#' be used to set `SPECIESPLUS_TOKEN` for the current session.
 #' @param verbose a logical. Should extra information be reported on progress?
 #'
 #' @return
-#' To be added 
+#' If `raw=TRUE`, then a object of class `spp_raw` is returned, which is essentially
+#' a list of lists. If `raw=FALSE`, then a object of class `spp_taxon` is returned,
+#' it is a collection of seven data frames:
+#' 1. `all_id`: general information for all entries, including non-active taxon concepts,
+#' 2. `general`: includes general information for active taxon concepts,
+#' 3. `higher_taxa`: includes taxonomy information,
+#' 4. `accepted_names`: list of accepted names (only for synonyms),
+#' 5. `common_names`: list of common names (only for accepted names),
+#' 6. `synonyms`: list of synonyms (only for accepted names),
+#' 7. `cites_listing`: list of current CITES listings with annotations (missing if `taxonomy == 'CMS'`).
 #'
 #' @references
 #' \url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
@@ -37,93 +46,91 @@
 #' @examples
 #' \donttest{
 #' res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
-#' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', raw = TRUE)
-#' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
-#' res4 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = 1:5, language = 'EN')
-#' res5 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
+#' res2 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor', raw = TRUE)
+#' res3 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', pages = 1:5, language = 'EN')
+#' res4 <- spp_taxonconcept(query_taxon = '', pages = c(44))
 #' }
 
[…]
