### 2018-09-17 f9ac003
rebuild the vignettes
