+++ inst/doc/elephant.Rmd
@@ -51,25 +51,22 @@ Text about taxonconcept outcomes to come ...
 
 Before taking more insights into the legislation status of the elephant, we have a look at where the elephant actually occurs naturally. For this, we can access the elephant's distribution information with the `spp_distributions()` function. Thereafter, we can map the distribution with the help of the `rworldmap` package.
 
-```{r map, message = FALSE, fig.width = 6.4, fig.height = 7.6, fig.align="center"}
-if("rworldmap" %in% rownames(installed.packages()) == FALSE){
-  install.packages("rworldmap")} # installing `rworldmap` if not installed yet
-suppressPackageStartupMessages(library(rworldmap))
+```{r map, eval = FALSE, message = FALSE, fig.width = 6.4, fig.height = 7.6, fig.align="center"}
+library(rworldmap)
 
 par(las = 1)
-map1 <- as.data.frame(spp_distributions(taxon_id = "4521"))
-map2 <- joinCountryData2Map(map1,
+ele_distr <- spp_distributions(taxon_id = "4521")$distributions
+map2 <- joinCountryData2Map(ele_distr,
                             joinCode="ISO2",
                             nameJoinColumn = "iso_code2",
                             nameCountryColumn = "name")
-map2@data$iso2 <- unlist(lapply(map2$iso_code2, is.null))
 plot(c(-23, 62), c(45, -40),
      type = "n",
      main = "Loxodonta africana",
      xlab = "Longitude",
      ylab = "Latitude")
-plot(map2, add = T)
-plot(map2[!map2$iso2,], col = "#bda25d", add = T)
+plot(map2, add = TRUE)
+plot(map2[!is.na(map2$iso_code2),], col = "#bda25d", add = TRUE)
 ```
 
 
@@ -81,6 +78,7 @@ First, we have a look at the CITES legislation status:
 
 ```{r}
 elephant_cites <- spp_cites_legislation(taxon_id = "4521")
+elephant_cites
 ```
 
 Text about CITES legislation outcomes to come ...
@@ -95,8 +93,7 @@ Can we make some nice map about trade directions (as a combination of map2 and t
 Last but not least, especially scientists will probably be interested in what references the Speciesplus information about the elephant is based on. For this, we can access the Speciesplus reference data with the `spp_references()` function. Within the function, we can select to access either the taxonomic or distribution references with the `type` argument.
 
 ```{r}
-elephant_ref_tax <- spp_references(taxon_id = "4521")
-elephant_ref_distr <- spp_references(taxon_id = "4521")
+elephant_refs <- spp_references(taxon_id = "4521")
 ```
 
 Text about references outcomes to come ...
