### 2018-09-21 a2bfd35
simplify listings more efficiently
