+++ man/spp_taxonconcept.Rd
@@ -63,7 +63,7 @@ Retrieve the taxon concept of a specific taxon (scientific name).
 \donttest{
 res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
 res2 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor', raw = TRUE)
-res3 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', pages = 1:5, language = 'EN')
+res3 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', pages = 1:3, language = 'EN')
 res4 <- spp_taxonconcept(query_taxon = '', pages = c(44))
 }
 }
