+++ man/spp_eu_legislation.Rd
@@ -44,8 +44,9 @@ given taxon concept (identifier must be known).
 \examples{
 \donttest{
 res1 <- spp_eu_legislation(taxon_id = '4521')
-res2 <- spp_eu_legislation(taxon_id = '4521', scope = 'historic')
-res3 <- spp_eu_legislation(taxon_id = '4521', scope = 'all', language='fr')
+res2 <- spp_eu_legislation(taxon_id = '3210')
+res3 <- spp_eu_legislation(taxon_id = '4521', scope = 'historic')
+res4 <- spp_eu_legislation(taxon_id = '4521', scope = 'all', language='fr')
 }
 }
 \references{
