### 2018-09-21 c4265f9
simplify outputs in a more efficient way
