+++ R/print.R
@@ -23,20 +23,20 @@ print.spp_raw <- function(x, ...) {
 #' @rdname print.spp
 #' @method print spp_taxon
 print.spp_taxon <- function(x, ...) {
-
+    
     rcites_print_title("General info:", "\n")
     rcites_print_df(x$general)
-
+    
     rcites_print_title("Classification:", "\n", "\n")
     rcites_print_df(x$higher_taxa)
-
+    
     rcites_print_title("Synonyms:", "\n", "\n")
     rcites_print_df(x$synonyms)
-
+    
     rcites_print_title("Common names:", "\n", "\n")
     rcites_print_df(x$common_names)
-
-    cat("\nInformation available:", paste(paste0("$", names(x)), collapse = ", "),
+    
+    cat("\nInformation available:", paste(paste0("$", names(x)), collapse = ", "), 
         "\n")
 }
 
@@ -44,10 +44,10 @@ print.spp_taxon <- function(x, ...) {
 #' @method print spp_eu_leg
 #' @rdname print.spp
 print.spp_eu_leg <- function(x, ...) {
-
+    
     rcites_print_title("EU listings ($eu_listings):", "\n")
     rcites_print_df_rm(x$eu_listings, col_rm = c("annotation", "hash_annotation"))
-
+    
     rcites_print_title("EU decisions ($eu_decisions):", "\n", "\n")
     npr <- c("notes", "eu_decision_type.description", "start_event.url")
     rcites_print_df_rm(x$eu_decisions, col_rm = npr)
@@ -57,16 +57,16 @@ print.spp_eu_leg <- function(x, ...) {
 #' @method print spp_cites_leg
 #' @rdname print.spp
 print.spp_cites_leg <- function(x, ...) {
-
+    
     rcites_print_title("Cites listings ($cites_listings):", "\n")
     rcites_print_df_rm(x$cites_listings, col_rm = c("annotation", "hash_annotation"))
-
+    
     rcites_print_title("Cites quotas ($cites_quotas):", "\n", "\n")
     rcites_print_df_rm(x$cites_quotas, col_rm = c("notes", "url"))
-
-    rcites_print_title("Cites suspensions ($cites_suspensions):", "\n",
+    
+    rcites_print_title("Cites suspensions ($cites_suspensions):", "\n", 
         "\n")
-    rcites_print_df_rm(x$cites_suspensions, col_rm = c("notes", paste0("start_notification.",
+    rcites_print_df_rm(x$cites_suspensions, col_rm = c("notes", paste0("start_notification.", 
         c("name", "date", "url"))))
 }
 
@@ -84,7 +84,7 @@ print.spp_refs <- function(x, ...) {
 print.spp_distr <- function(x, ...) {
     rcites_print_title("Distributions ($distributions):", "\n")
     rcites_print_df(x$distributions)
-
+    
     rcites_print_title("References ($references):", "\n", "\n")
     rcites_print_df_rm(x$references)
 }
