### 2018-09-22 33a6642
test rcites_to_logical() :gear:
