+++ R/zzz.R
@@ -39,14 +39,14 @@ rcites_timestamp <- function(x) {
 
 rcites_lang <- function(x) {
     out <- match.arg(x, c("en", "fr", "es"))
-    if (out == "en") 
+    if (out == "en")
         out <- NULL else out <- paste0("language=", out)
     out
 }
 
 rcites_scope <- function(x) {
     out <- match.arg(x, c("current", "historic", "all"))
-    if (out == "current") 
+    if (out == "current")
         out <- NULL else out <- paste0("scope=", out)
     out
 }
@@ -83,17 +83,17 @@ rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 ## Pagination helpers
 
 # auto pagination
-rcites_autopagination <- function(q_url, per_page, pages, tot_page, token, 
+rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
     verbose = TRUE) {
     out <- list()
-    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$", 
+    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$",
         replacement = "")
     # grepl('\\.json?', pat = '\\.json\\?$')
     for (i in seq_along(pages)) {
-        if (verbose) 
-            cat("Retrieving info from page ", pages[i], "/", tot_page, 
+        if (verbose)
+            cat("Retrieving info from page ", pages[i], "/", tot_page,
                 "\r")
-        q_url_new <- paste0(q_url_0, "page=", pages[i], "&per_page=", min(per_page, 
+        q_url_new <- paste0(q_url_0, "page=", pages[i], "&per_page=", min(per_page,
             500))
         out[[i]] <- rcites_res(q_url_new, token)
     }
@@ -103,7 +103,7 @@ rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
     out
 }
 
-# 
+#
 rcites_numberpages <- function(x) {
     x$total_entries%/%x$per_page + (x$total_entries%%x$per_page > 0)
 }
@@ -131,7 +131,7 @@ rcites_null_to_na <- function(x) {
 }
 
 # convert certain columns to logical
-rcites_to_logical <- function(x, id) {
+rcites_to_logical <- function(x) {
     vc_nm <- c("is_current", "applies_to_import", "public_display")
     id <- which(names(x) %in% vc_nm)
     if (length(id)) {
@@ -153,18 +153,18 @@ rcites_assign_class <- function(x) {
 }
 
 
-# 
+#
 rcites_simplify_listings <- function(x) {
     # fields below may or may not be included, so there are removed
     vc_rm <- c("party", "hash_annotation", "annotation")
-    tmp <- lapply(lapply(x, FUN = function(x) x[!names(x) %in% vc_rm]), 
+    tmp <- lapply(lapply(x, FUN = function(x) x[!names(x) %in% vc_rm]),
         FUN = function(y) data.frame(do.call(cbind, y)))
     if (length(tmp) > 1) {
         out <- do.call(rbind, tmp)
     } else {
         out <- tmp[[1L]]
     }
-    # 
+    #
     out <- rcites_to_logical(out)
     out <- rcites_assign_class(out)
     out
@@ -172,32 +172,32 @@ rcites_simplify_listings <- function(x) {
 
 
 
-# 
+#
 rcites_simplify_decisions <- function(x) {
     # these fields may or may not be included so I removed them
     tmp0 <- lapply(lapply(x, rcites_null_to_na), unlist)
[…]
