### 2018-09-23 39b9ca0
add r_cites_shorten :gear:
