### 2018-09-23 2560511
test rcites_print_shorten() :shield:
