+++ man/spp_references.Rd
@@ -34,7 +34,8 @@ Retrieve available references for a given taxon concept.
 \examples{
 \donttest{
 res1 <- spp_references(taxon_id = '4521')
-res2 <- spp_references(taxon_id = '4521', raw = TRUE)
+res2 <- spp_references(taxon_id = '3210')
+res3 <- spp_references(taxon_id = '4521', raw = TRUE)
 }
 }
 \references{
