### 2018-10-02 e6102e0
few readme edits :pencil:
