+++ R/zzz.R
@@ -8,9 +8,9 @@
 #' @keywords internal
 "_PACKAGE"
 
-# Helper functions
+# HELPER FUNCTIONS
 
-
+##################
 ## General helpers
 
 rcites_baseurl <- function() "https://api.speciesplus.net/api/v1/"
@@ -55,7 +55,7 @@ rcites_scope <- function(x) {
 rcites_checkid <- function(taxon_id) {
     # id check
     if (!grepl(taxon_id, pattern = "^[0-9]*$")) {
-        stop("The taxon concept identifier is made of digits only.")
+        stop("A taxon concept identifier is made up of digits only.")
     }
     invisible(NULL)
 }
@@ -81,7 +81,7 @@ rcites_getsecret <- function() {
 rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 
 
-####################
+#####################
 ## Pagination helpers
 
 rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
@@ -105,7 +105,7 @@ rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
 }
 
 rcites_numberpages <- function(x) {
-    x$total_entries%/%x$per_page + (x$total_entries%%x$per_page > 0)
+    x$total_entries %/% x$per_page + (x$total_entries %% x$per_page > 0)
 }
 
 
@@ -121,7 +121,7 @@ rcites_add_author_year <- function(x) {
     x
 }
 
-# convert null to na
+# convert null to na recursively
 rcites_null_to_na <- function(x) {
     if (is.list(x)) {
         return(lapply(x, rcites_null_to_na))
@@ -173,7 +173,6 @@ rcites_simplify_listings <- function(x) {
 
 #
 rcites_simplify_decisions <- function(x) {
-    # these fields may or may not be included so I removed them
     tmp0 <- lapply(lapply(x, rcites_null_to_na), unlist)
     out <- data.frame(do.call(rbind, lapply(tmp0, function(y) data.frame(rbind(y)))))
     #
@@ -184,10 +183,11 @@ rcites_simplify_decisions <- function(x) {
 
 #
 rcites_simplify_distributions <- function(x) {
-    # these fields may or may not be included so I removed them
     tmp <- lapply(x, rcites_null_to_na)
     out <- list()
-    out$distributions <- data.frame(do.call(rbind, lapply(tmp, function(y) data.frame(rbind(unlist(y[!names(y) %in%
+    out$distributions <- data.frame(do.call(rbind, lapply(tmp, function(y)
+    # these fields may or may not be included so I removed them
+    data.frame(rbind(unlist(y[!names(y) %in%
         c("tags", "references")]))))))
     out$distributions$tags <- unlist(lapply(tmp, function(y) paste(y$tags,
         collapse = ", ")))
@@ -203,6 +203,7 @@ rcites_simplify_distributions <- function(x) {
 
 
 
+################
 ## print helpers
 
 rcites_print_shorten <- function(x, stop = 20) {
@@ -218,7 +219,7 @@ rcites_print_title <- function(x, after = "", before = "") {
     cat(before, x, "\n", paste(rep("-", nchar(x)), collapse = ""), after,
         sep = "")
 }
-#
+
 rcites_print_df <- function(x, nrows = 10) {
     if ("tibble" %in% .packages()) {
         # tibble truncates the outputs already
@@ -241,36 +242,37 @@ rcites_print_df_rm <- function(x, col_rm = "", nrows = 10) {
 }
 
 
+#############################
 ## spp_taxonconcept() helpers
 
[…]
