### 2018-10-07 92d008a
simplify higher_taxa
