+++ R/zzz.R
@@ -8,7 +8,7 @@
 #' @keywords internal
 "_PACKAGE"
 
-# HELPER FUNCTIONS
+# Helper functions
 
 ##################
 ## General helpers
@@ -55,14 +55,15 @@ rcites_scope <- function(x) {
 rcites_checkid <- function(taxon_id) {
     # id check
     if (!grepl(taxon_id, pattern = "^[0-9]*$")) {
-        stop("A taxon concept identifier is made up of digits only.")
+        stop("The taxon concept identifier is made of digits only.")
     }
     invisible(NULL)
 }
 
 
-####################
-## Secret helpers ##
+
+#################
+## Secret helpers
 
 # See https://cran.r-project.org/web/packes/httr/vignettes/secrets.html
 rcites_getsecret <- function() {
@@ -105,7 +106,7 @@ rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
 }
 
 rcites_numberpages <- function(x) {
-    x$total_entries %/% x$per_page + (x$total_entries %% x$per_page > 0)
+    x$total_entries%/%x$per_page + (x$total_entries%%x$per_page > 0)
 }
 
 
@@ -121,7 +122,7 @@ rcites_add_author_year <- function(x) {
     x
 }
 
-# convert null to na recursively
+# convert null to na
 rcites_null_to_na <- function(x) {
     if (is.list(x)) {
         return(lapply(x, rcites_null_to_na))
@@ -152,7 +153,6 @@ rcites_assign_class <- function(x) {
 }
 
 
-#
 rcites_simplify_listings <- function(x) {
     # fields below may or may not be included, so there are removed
     vc_rm <- c("party", "hash_annotation", "annotation")
@@ -169,9 +169,6 @@ rcites_simplify_listings <- function(x) {
     out
 }
 
-
-
-#
 rcites_simplify_decisions <- function(x) {
     tmp0 <- lapply(lapply(x, rcites_null_to_na), unlist)
     out <- data.frame(do.call(rbind, lapply(tmp0, function(y) data.frame(rbind(y)))))
@@ -181,14 +178,13 @@ rcites_simplify_decisions <- function(x) {
     out
 }
 
-#
 rcites_simplify_distributions <- function(x) {
     tmp <- lapply(x, rcites_null_to_na)
     out <- list()
-    out$distributions <- data.frame(do.call(rbind, lapply(tmp, function(y)
     # these fields may or may not be included so I removed them
-    data.frame(rbind(unlist(y[!names(y) %in%
+    out$distributions <- data.frame(do.call(rbind, lapply(tmp, function(y) data.frame(rbind(unlist(y[!names(y) %in%
         c("tags", "references")]))))))
+    # collapse tags
     out$distributions$tags <- unlist(lapply(tmp, function(y) paste(y$tags,
         collapse = ", ")))
     out$distributions <- rcites_assign_class(out$distributions)
@@ -203,7 +199,7 @@ rcites_simplify_distributions <- function(x) {
 
 
 
-################
+####################
 ## print helpers
 
 rcites_print_shorten <- function(x, stop = 20) {
@@ -219,7 +215,7 @@ rcites_print_title <- function(x, after = "", before = "") {
     cat(before, x, "\n", paste(rep("-", nchar(x)), collapse = ""), after,
         sep = "")
 }
-
+#
 rcites_print_df <- function(x, nrows = 10) {
     if ("tibble" %in% .packages()) {
         # tibble truncates the outputs already
@@ -242,47 +238,44 @@ rcites_print_df_rm <- function(x, col_rm = "", nrows = 10) {
 }
 
 
+
[…]
