### 2018-10-07 2e11b3a
simplify outputs
