+++ R/spp_taxonconcept.R
@@ -53,39 +53,40 @@
 #' }
 
 
-spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE,
-    language = NULL, updated_since = NULL, per_page = 500, pages = NULL,
+spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants = FALSE, 
+    language = NULL, updated_since = NULL, per_page = 500, pages = NULL, 
     raw = FALSE, token = NULL, verbose = TRUE) {
     # taxonomy check
     taxonomy <- match.arg(taxonomy, c("CITES", "CMS"))
     # token check
-    if (is.null(token))
+    if (is.null(token)) 
         token <- rcites_getsecret()
     # request
     if (is.null(pages)) {
+        # first page
         f_page <- 1
     } else {
         pages <- sort(unique(as.integer(pages)))
-        f_page <- unique(pages[1L])
+        f_page <- pages[1L]
     }
-    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy,
+    q_url <- rcites_taxonconcept_request(query_taxon, token, taxonomy, 
         with_descendants, f_page, per_page, updated_since, language)
     # results
     tmp <- rcites_res(q_url, token)
     # number of pages
     pag <- rcites_numberpages(tmp$pagination)
-    #
+    # 
     if (!pag) {
         warning("Taxon not listed.")
         return(NULL)
     } else {
         if (pag > 1) {
-            if (is.null(pages))
-              pages <- seq_len(pag) else pages <- pages[pages <= pag]
-            if (!length(pages))
-              stop("Only page 1-", pag, " are available.")
+            if (is.null(pages)) 
+                pages <- seq_len(pag) else pages <- pages[pages <= pag]
+            if (!length(pages)) 
+                stop("Only page 1-", pag, " are available.")
             if (length(pages) > 1) {
-                res <- rcites_autopagination(q_url, per_page, pages[-1L],
+                res <- rcites_autopagination(q_url, per_page, pages[-1L], 
                   pag, token, verbose)
                 tmp2 <- c(tmp$taxon_concepts, do.call(c, lapply(res, function(x) x$taxon_concepts)))
             } else tmp2 <- tmp$taxon_concepts
@@ -96,42 +97,42 @@ spp_taxonconcept <- function(query_taxon, taxonomy = "CITES", with_descendants =
             return(tmp2)
         } else {
             ## special cases
-            sp_nm <- c("synonyms", "higher_taxa", "common_names", "cites_listing",
+            sp_nm <- c("synonyms", "higher_taxa", "common_names", "cites_listing", 
                 "cites_listings", "accepted_names")
             ## output
             out <- list()
             out$all_id <- rcites_taxonconcept_allentries(tmp2, sp_nm)
-            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at,
+            out$all_id$updated_at <- as.POSIXlt(out$all_id$updated_at, 
                 format = "%Y-%m-%dT%H:%M:%OS")
             out$all_id$active <- as.logical(out$all_id$active)
             ## extract active only
             id <- which(out$all_id$active)
             out$general <- out$all_id[out$all_id$active, ]
-
+            
             ## Classification
-            out$higher_taxa <- rcites_taxonconcept_higher_taxa(tmp2[id],
+            out$higher_taxa <- rcites_taxonconcept_higher_taxa(tmp2[id], 
                 out$general$id)
             ## Names
-            out$accepted_names <- rcites_taxonconcept_special_cases(tmp2[!id],
-                name = "accepted_names", out$all_id$id[!id])
-            out$common_names <- rcites_taxonconcept_special_cases(tmp2[id],
[…]
