### 2018-10-07 e0e2648
warning not error for checkid
