### 2018-10-07 c196776
create rcites_combine_list() :hammer:
