+++ R/spp_distributions.R
@@ -30,22 +30,28 @@
 #' @examples
 #' \donttest{
 #'  res1 <- spp_distributions(taxon_id = '4521')
-#'  res2 <- spp_distributions(taxon_id = '3210')
+#'  res2 <- spp_distributions(taxon_id = c('4521', '3210'))
 #'  res3 <- spp_distributions(taxon_id = '4521', raw = TRUE)
 #'  res4 <- spp_distributions(taxon_id = '4521', language = 'fr')
 #' }
 
 spp_distributions <- function(taxon_id, language = "en", raw = FALSE, token = NULL) {
+
+    if (length(taxon_id)>1) {
+      out <- lapply(taxon_id, spp_distributions, language = "en", raw = raw, token = token)
+      do.call(c, cool)
+    } else {
     # token check
-    if (is.null(token)) 
+    if (is.null(token))
         token <- rcites_getsecret()
     # id check
-    rcites_checkid(taxon_id)
+    if (rcites_checkid(taxon_id))
+      return(NULL)
     # set query_string
     tmp <- rcites_lang(language)
-    if (!is.null(tmp)) 
+    if (!is.null(tmp))
         tmp <- paste0("?", tmp)
-    q_url <- rcites_url("taxon_concepts/", taxon_id, "/distributions.json", 
+    q_url <- rcites_url("taxon_concepts/", taxon_id, "/distributions.json",
         tmp)
     # get results
     res <- rcites_res(q_url, token)
@@ -57,5 +63,6 @@ spp_distributions <- function(taxon_id, language = "en", raw = FALSE, token = NU
         out <- rcites_simplify_distributions(res)
         class(out) <- c("spp_distr")
     }
+  }
     out
 }
