### 2018-10-08 53e4c8e
vectorize spp_distributions()
