+++ R/zzz.R
@@ -39,14 +39,14 @@ rcites_timestamp <- function(x) {
 
 rcites_lang <- function(x) {
     out <- match.arg(x, c("en", "fr", "es"))
-    if (out == "en")
+    if (out == "en") 
         out <- NULL else out <- paste0("language=", out)
     out
 }
 
 rcites_scope <- function(x) {
     out <- match.arg(x, c("current", "historic", "all"))
-    if (out == "current")
+    if (out == "current") 
         out <- NULL else out <- paste0("scope=", out)
     out
 }
@@ -62,33 +62,39 @@ rcites_checkid <- function(taxon_id) {
 }
 
 rcites_current_id <- function(x) {
-  cat(">>> Now processing taxon_id '", x, "'............", sep = "")
+    cat(">>> Now processing taxon_id '", x, "'............", sep = "")
 }
 
 
-rcites_combine_lists <- function(x, taxon_id) {
-  wch <- !unlist(lapply(x, is.null))
-  x <- Filter(Negate(is.null), x)
-  cls <- class(x[[1L]])
-  # get names
-  ls_keys <- lapply(x, names)
-  if (!all(unlist(lapply(ls_keys, identical, ls_keys[[1L]])))) {
-      stop("Cannot combine lists with different names")
-  }
-  #
-  for (i in seq_along(x)) {
-    x[[i]] <- lapply(x[[i]], function(y) cbind(taxon_id = taxon_id[wch][i], y))
-  }
-  #
-  out <- list()
-  for (i in seq_along(ls_keys[[1L]])) {
-    key <- ls_keys[[1L]][i]
-    out[[key]] <- do.call(rbind, lapply(x, `[[`, key))
-  }
-  class(out) <- paste0(cls, "_multi", sep = "")
-  out
+rcites_combine_lists <- function(x, taxon_id, raw) {
+    wch <- !unlist(lapply(x, is.null))
+    x <- Filter(Negate(is.null), x)
+    cls <- class(x[[1L]])
+    if (raw) {
+        out <- x
+        out$taxon_ids <- taxon_id[wch]
+        class(out) <- c("list", "spp_raw_multi")
+    } else {
+        # get names
+        ls_keys <- lapply(x, names)
+        if (!all(unlist(lapply(ls_keys, identical, ls_keys[[1L]])))) {
+            stop("Cannot combine lists with different names")
+        }
+        # 
+        for (i in seq_along(x)) {
+            x[[i]] <- lapply(x[[i]], function(y) cbind(taxon_id = taxon_id[wch][i], 
+                y))
+        }
+        # 
+        out <- list()
+        for (i in seq_along(ls_keys[[1L]])) {
+            key <- ls_keys[[1L]][i]
+            out[[key]] <- do.call(rbind, lapply(x, `[[`, key))
+        }
+        class(out) <- paste0(cls, "_multi", sep = "")
+    }
+    out
 }
-
 
 
 
@@ -113,23 +119,23 @@ rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 
 ##################### Pagination helpers
 
-rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
+rcites_autopagination <- function(q_url, per_page, pages, tot_page, token, 
     verbose = TRUE) {
     out <- list()
-    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$",
+    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$", 
         replacement = "")
-    #
+    # 
     for (i in seq_along(pages)) {
-        if (verbose)
[…]
