+++ man/spp_taxonconcept.Rd
@@ -65,7 +65,7 @@ Retrieve the taxon concept of a specific taxon (scientific name).
 res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
 res2 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor', raw = TRUE)
 res3 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', pages = c(1,3), language = 'EN')
-res4 <- spp_taxonconcept(query_taxon = '', pages = 44)
+res4 <- spp_taxonconcept(query_taxon = '', per_page = 20, pages = 44)
 }
 }
 \references{
