+++ man/print.spp.Rd
@@ -3,14 +3,18 @@
 \name{print.spp}
 \alias{print.spp}
 \alias{print.spp_raw}
+\alias{print.spp_raw_multi}
 \alias{print.spp_taxon}
 \alias{print.spp_eu_leg}
 \alias{print.spp_cites_leg}
 \alias{print.spp_refs}
 \alias{print.spp_distr}
+\alias{print.spp_distr_multi}
 \title{Print methods for objects of class \code{spp_*}}
 \usage{
 \method{print}{spp_raw}(x, ...)
+
+\method{print}{spp_raw_multi}(x, ...)
 
 \method{print}{spp_taxon}(x, ...)
 
@@ -20,7 +24,9 @@
 
 \method{print}{spp_refs}(x, ...)
 
-\method{print}{spp_refs}(x, ...)
+\method{print}{spp_distr}(x, ...)
+
+\method{print}{spp_distr_multi}(x, ...)
 }
 \arguments{
 \item{x}{The result object.}

[…]
