### 2018-10-08 024e94c
verbose argument in ref and legislation functions
