+++ R/spp_references.R
@@ -10,6 +10,7 @@
 #' `NULL` and requires the environment variable `SPECIESPLUS_TOKEN` to be
 #' set directly in `Renviron`. Alternatively, [set_token()] can be used to set
 #' `SPECIESPLUS_TOKEN` for the current session.
+#' @param verbose a logical. Should extra information be reported on progress?
 #'
 #' @return If `raw` is set to `TRUE` then an object of class `spp_raw` is returned
 #' which is essentially the list of lists (see option `as = 'parsed'` in [httr::content()]).
@@ -30,7 +31,7 @@
 #' res3 <- spp_references(taxon_id = '4521', raw = TRUE)
 #' }
 
-spp_references <- function(taxon_id, raw = FALSE, token = NULL) {
+spp_references <- function(taxon_id, raw = FALSE, token = NULL, verbose = TRUE) {
     # token check
     if (is.null(token)) 
         token <- rcites_getsecret()
