+++ man/spp_references.Rd
@@ -4,7 +4,7 @@
 \alias{spp_references}
 \title{Get references for a given taxon concept}
 \usage{
-spp_references(taxon_id, raw = FALSE, token = NULL)
+spp_references(taxon_id, raw = FALSE, token = NULL, verbose = TRUE)
 }
 \arguments{
 \item{taxon_id}{character string containing a species' taxon concept identifier
@@ -17,6 +17,8 @@ spp_references(taxon_id, raw = FALSE, token = NULL)
 \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
 set directly in \code{Renviron}. Alternatively, \code{\link[=set_token]{set_token()}} can be used to set
 \code{SPECIESPLUS_TOKEN} for the current session.}
+
+\item{verbose}{a logical. Should extra information be reported on progress?}
 }
 \value{
 If \code{raw} is set to \code{TRUE} then an object of class \code{spp_raw} is returned
