### 2018-10-08 10c4c62
vectorize spp_eu_legislation
