+++ man/print.spp.Rd
@@ -6,6 +6,7 @@
 \alias{print.spp_raw_multi}
 \alias{print.spp_taxon}
 \alias{print.spp_eu_leg}
+\alias{print.spp_eu_leg_multi}
 \alias{print.spp_cites_leg}
 \alias{print.spp_refs}
 \alias{print.spp_distr}
@@ -19,6 +20,8 @@
 \method{print}{spp_taxon}(x, ...)
 
 \method{print}{spp_eu_leg}(x, ...)
+
+\method{print}{spp_eu_leg_multi}(x, ...)
 
 \method{print}{spp_cites_leg}(x, ...)
 
