+++ man/spp_eu_legislation.Rd
@@ -46,7 +46,7 @@ given taxon concept (identifier must be known).
 \examples{
 \donttest{
 res1 <- spp_eu_legislation(taxon_id = '4521')
-res2 <- spp_eu_legislation(taxon_id = '3210')
+res2 <- spp_eu_legislation(taxon_id = c('4521', '3210', '10255'))
 res3 <- spp_eu_legislation(taxon_id = '4521', scope = 'historic')
 res4 <- spp_eu_legislation(taxon_id = '4521', scope = 'all', language='fr')
 }
