### 2018-10-08 96db055
vectorize spp_cites_legislation()
