+++ R/print.R
@@ -37,17 +37,17 @@ print.spp_taxon <- function(x, ...) {
     gen <- paste0("General info (", attributes(x)$taxonomy, "):")
     rcites_print_title(gen, "\n")
     rcites_print_df(x$general)
-    
+
     rcites_print_title("Classification:", "\n", "\n")
     rcites_print_df(x$higher_taxa)
-    
+
     rcites_print_title("Synonyms:", "\n", "\n")
     rcites_print_df(x$synonyms)
-    
+
     rcites_print_title("Common names:", "\n", "\n")
     rcites_print_df(x$common_names)
-    
-    cat("\nInformation available:", paste(paste0("$", names(x)), collapse = ", "), 
+
+    cat("\nInformation available:", paste(paste0("$", names(x)), collapse = ", "),
         "\n")
 }
 
@@ -55,10 +55,10 @@ print.spp_taxon <- function(x, ...) {
 #' @method print spp_eu_leg
 #' @rdname print.spp
 print.spp_eu_leg <- function(x, ...) {
-    
+
     rcites_print_title("EU listings ($eu_listings):", "\n")
     rcites_print_df_rm(x$eu_listings, col_rm = c("annotation", "hash_annotation"))
-    
+
     rcites_print_title("EU decisions ($eu_decisions):", "\n", "\n")
     npr <- c("notes", "eu_decision_type.description", "start_event.url")
     rcites_print_df_rm(x$eu_decisions, col_rm = npr)
@@ -76,17 +76,25 @@ print.spp_eu_leg_multi <- function(x, ...) {
 #' @method print spp_cites_leg
 #' @rdname print.spp
 print.spp_cites_leg <- function(x, ...) {
-    
+
     rcites_print_title("Cites listings ($cites_listings):", "\n")
     rcites_print_df_rm(x$cites_listings, col_rm = c("annotation", "hash_annotation"))
-    
+
     rcites_print_title("Cites quotas ($cites_quotas):", "\n", "\n")
     rcites_print_df_rm(x$cites_quotas, col_rm = c("notes", "url"))
-    
-    rcites_print_title("Cites suspensions ($cites_suspensions):", "\n", 
+
+    rcites_print_title("Cites suspensions ($cites_suspensions):", "\n",
         "\n")
-    rcites_print_df_rm(x$cites_suspensions, col_rm = c("notes", paste0("start_notification.", 
+    rcites_print_df_rm(x$cites_suspensions, col_rm = c("notes", paste0("start_notification.",
         c("name", "date", "url"))))
+}
+
+#' @method print spp_eu_leg_multi
+#' @rdname print.spp
+print.spp_cites_leg_multi <- function(x, ...) {
+    rcites_print_taxon_id(x$cites_listings$taxon_id)
+    cat("\n")
+    print.spp_cites_leg(x)
 }
 
 
@@ -94,7 +102,7 @@ print.spp_cites_leg <- function(x, ...) {
 #' @rdname print.spp
 print.spp_refs <- function(x, ...) {
     rcites_print_title("References ($references):", "\n")
-    x$references$citation <- rcites_print_shorten(x$references$citation, 
+    x$references$citation <- rcites_print_shorten(x$references$citation,
         60)
     rcites_print_df(x$references)
 }
@@ -105,7 +113,7 @@ print.spp_refs <- function(x, ...) {
 print.spp_distr <- function(x, ...) {
     rcites_print_title("Distributions ($distributions):", "\n")
     rcites_print_df(x$distributions)
-    
+
     rcites_print_title("References ($references):", "\n", "\n")
     x$references$reference <- rcites_print_shorten(x$references$reference)
     rcites_print_df_rm(x$references)
