### 2018-10-08 994e25f
vetorize spp_references()
