+++ R/spp_references.R
@@ -27,29 +27,41 @@
 #' @examples
 #' \donttest{
 #' res1 <- spp_references(taxon_id = '4521')
-#' res2 <- spp_references(taxon_id = '3210')
+#' res2 <- spp_references(c('4521', '3210', '10255'))
 #' res3 <- spp_references(taxon_id = '4521', raw = TRUE)
 #' }
 
 spp_references <- function(taxon_id, raw = FALSE, token = NULL, verbose = TRUE) {
-    # token check
-    if (is.null(token)) 
-        token <- rcites_getsecret()
-    # id check
-    rcites_checkid(taxon_id)
-    ## create url
-    q_url <- rcites_url("taxon_concepts/", taxon_id, "/references.json")
-    ## get_res
-    tmp <- rcites_res(q_url, token)
-    ## outputs
-    if (raw) {
-        out <- tmp
-        class(out) <- c("list", "spp_raw")
-    } else {
-        out <- list()
-        out$references <- rcites_simplify_decisions(tmp)
-        class(out) <- c("spp_refs")
-    }
-    # 
-    out
+
+  if (length(taxon_id) > 1) {
+      out <- lapply(taxon_id, spp_references, raw = raw, token = token, verbose = verbose)
+      out <- rcites_combine_lists(out, taxon_id, raw)
+  } else {
+      # token check
+      if (is.null(token))
+          token <- rcites_getsecret()
+      # id check
+      if (rcites_checkid(taxon_id)) {
+          out <- NULL
+      } else {
+          if (verbose)
+              rcites_current_id(taxon_id)
+              ## create url
+              q_url <- rcites_url("taxon_concepts/", taxon_id, "/references.json")
+              ## get_res
+              tmp <- rcites_res(q_url, token)
+              ## outputs
+              if (raw) {
+                  out <- tmp
+                  class(out) <- c("list", "spp_raw")
+              } else {
+                  out <- list()
+                  out$references <- rcites_simplify_decisions(tmp)
+                  class(out) <- c("spp_refs")
+              }
+          if (verbose)
+              cat(" done. \n")
+      }
+  }
+  out
 }
