+++ NEWS.md
@@ -2,16 +2,18 @@
 
 - new features:
 
-  - `spp_taxonconcept()` now includes an auto-pagination.
-  - Functions now returns S3 objects for which a print method is defined for all
-  these objects:
+  - `spp_taxonconcept()` now includes an auto-pagination that allows to retrieve
+  all entries for queries that have more than 500 results;
+  - Functions `spp_*` now returns S3 objects:
+    - `spp_taxonconcept()` returns an object of class `spp_taxon`;
+    - `spp_cites_legislation()` returns an object of class `spp_cites_leg`  or `spp_cites_leg_multi`;
+    - `spp_eu_legislation()` returns an object of class `spp_eu_leg` or `spp_eu_leg_multi`;
+    - `spp_distributions()` returns an object of class `spp_distr` or `spp_distr_multi`;
+    - `spp_references()` returns an object of class `spp_refs` or `spp_refs_multi`;
+  - Moreover:
     - `spp_raw` class is defined for all functions including the argument `raw`,
-    it returns the parsed output from the API as a list object.  
-    - `spp_taxonconcept` returns an object of class `spp_taxon`
-    - `spp_cites_legislation` returns an object of class `spp_cites_leg`
-    - `spp_eu_legislation` returns an object of class `spp_eu_leg`
-    - `spp_distributions` returns an object of class `spp_distr`
-    - `spp_references` returns an object of class `spp_refs`
+    it returns the parsed output from the API as a list object.   
+
 
 
 - add function parameters:
