+++ man/print.spp.Rd
@@ -9,9 +9,10 @@
 \alias{print.spp_eu_leg_multi}
 \alias{print.spp_cites_leg}
 \alias{print.spp_cites_leg_multi}
-\alias{print.spp_refs}
 \alias{print.spp_distr}
 \alias{print.spp_distr_multi}
+\alias{print.spp_refs}
+\alias{print.spp_refs_multi}
 \title{Print methods for objects of class \code{spp_*}}
 \usage{
 \method{print}{spp_raw}(x, ...)
@@ -28,11 +29,13 @@
 
 \method{print}{spp_eu_leg_multi}(x, ...)
 
-\method{print}{spp_refs}(x, ...)
-
 \method{print}{spp_distr}(x, ...)
 
 \method{print}{spp_distr_multi}(x, ...)
+
+\method{print}{spp_refs}(x, ...)
+
+\method{print}{spp_refs_multi}(x, ...)
 }
 \arguments{
 \item{x}{The result object.}
