+++ man/spp_references.Rd
@@ -36,7 +36,7 @@ Retrieve available references for a given taxon concept.
 \examples{
 \donttest{
 res1 <- spp_references(taxon_id = '4521')
-res2 <- spp_references(taxon_id = '3210')
+res2 <- spp_references(c('4521', '3210', '10255'))
 res3 <- spp_references(taxon_id = '4521', raw = TRUE)
 }
 }
