+++ tests/testthat/test-legislations.R
@@ -2,18 +2,20 @@ context("Legislations")
 
 skip_on_cran()
 skip_if_no_auth()
-#
-res1 <- spp_cites_legislation(taxon_id = tx_id)
-res1b <- spp_cites_legislation(taxon_id = tx_id2)
-res2 <- spp_cites_legislation(taxon_id = tx_id, raw = TRUE)
 ut_pause()
-res3 <- spp_cites_legislation(taxon_id = tx_id, scope = 'all')
-res4 <- spp_cites_legislation(taxon_id = tx_id, language = 'fr')
+
+#
+res1 <- spp_cites_legislation(taxon_id = tx_id, verbose = FALSE)
+res1b <- spp_cites_legislation(taxon_id = tx_id2, verbose = FALSE)
+res2 <- spp_cites_legislation(taxon_id = tx_id, raw = TRUE, verbose = FALSE)
+ut_pause()
+res3 <- spp_cites_legislation(taxon_id = tx_id, scope = 'all', verbose = FALSE)
+res4 <- spp_cites_legislation(taxon_id = tx_id, language = 'fr', verbose = FALSE)
 ut_pause()
 #
-res5 <- spp_eu_legislation(taxon_id = tx_id)
-res5b <- spp_eu_legislation(taxon_id = tx_id2)
-res6 <- spp_eu_legislation(taxon_id = tx_id, raw = TRUE)
+res5 <- spp_eu_legislation(taxon_id = tx_id, verbose = FALSE)
+res5b <- spp_eu_legislation(taxon_id = tx_id2, verbose = FALSE)
+res6 <- spp_eu_legislation(taxon_id = tx_id, raw = TRUE, verbose = FALSE)
 ut_pause()
 res7 <- spp_eu_legislation(taxon_id = tx_id, scope = 'all')
 res8 <- spp_eu_legislation(taxon_id = tx_id, language = 'fr')
@@ -69,3 +71,38 @@ test_that("Language", {
   expect_true(all(lang_en_eu == c(TRUE, FALSE)))
   expect_true(all(lang_fr_eu == c(FALSE, TRUE)))
 })
+
+
+ut_pause(1)
+res9 <- spp_cites_legislation(taxon_id = c('4521', '3210'))
+ut_pause()
+res10 <- spp_eu_legislation(taxon_id = c('4521', '3210'))
+ut_pause()
+res11 <- spp_cites_legislation(taxon_id = c('4521', '3210'), raw = TRUE, verbose = FALSE)
+ut_pause()
+res12 <- spp_eu_legislation(taxon_id = c('4521', '3210'), raw = TRUE, verbose = FALSE)
+
+
+test_that("leg_multi outputs", {
+  expect_equal(class(res9), "spp_cites_leg_multi")
+  expect_equal(class(res10), "spp_eu_leg_multi")
+  #
+  expect_equal(nrow(res9$cites_listings), nrow(res1$cites_listings) + nrow(res1b$cites_listings))
+  expect_equal(nrow(res10$eu_listings), nrow(res5$eu_listings) + nrow(res5b$eu_listings))
+  #
+  expect_identical(unique(res9$cites_listings$taxon_id),  c('4521', '3210'))
+  expect_identical(unique(res10$eu_listings$taxon_id),  c('4521', '3210'))
+  #
+  expect_identical(class(res11), cl_raw_multi)
+  expect_identical(class(res12), cl_raw_multi)
+  #
+  expect_equal(class(res11[[1L]]), cl_raw)
+  expect_equal(class(res12[[1L]]), cl_raw)
+  #
+  expect_equal(length(res11), 3)
+  expect_equal(length(res12), 3)
+  #
[…]
