### 2018-10-08 6f184f5
Vectorize (#36)
* better comment spp_taxonconcept

* simplify higher_taxa

* simplify outputs

* warning not error for checkid

* create rcites_combine_list() :hammer:

* vectorize spp_distributions()

* handle raw case

* verbose argument in ref and legislation functions

* vectorize spp_eu_legislation

* vectorize spp_cites_legislation()

* vetorize spp_references()

* rebuild website

* add tests :shield:
