+++ tests/testthat/test-pkghelpers.R
@@ -10,7 +10,7 @@ test_that("General helpers", {
   expect_true(is.null(rcites_scope("current")))
   expect_true(identical(rcites_scope("all"), "scope=all"))
   expect_error(rcites_scope("wrong"))
-  expect_error(rcites_checkid("1wrong234"))
+  expect_warning(rcites_checkid("1wrong234"))
 })
 
 
@@ -40,7 +40,7 @@ str <- paste(letters[1:26], collapse="")
 str2 <- paste(letters[1:25], collapse="")
 rcites_print_shorten(str)
 test_that("Print helpers", {
-  expect_equal(rcites_print_shorten(str), paste0(substring(str, 1, 20), "[...]"))
+  expect_equal(rcites_print_shorten(str), paste0(substring(str, 1, 20), " [truncated]"))
   expect_equal(rcites_print_shorten(str2), str2)
-  expect_equal(rcites_print_shorten(str, 4), "abcd[...]")
+  expect_equal(rcites_print_shorten(str, 4), "abcd [truncated]")
 })
