+++ R/zzz.R
@@ -8,9 +8,12 @@
 #' @keywords internal
 "_PACKAGE"
 
-# Helper functions
 
-################## General helpers
+# HELPER FUNCTIONS
+
+##################
+## General helpers
+
 
 rcites_baseurl <- function() "https://api.speciesplus.net/api/v1/"
 
@@ -39,14 +42,14 @@ rcites_timestamp <- function(x) {
 
 rcites_lang <- function(x) {
     out <- match.arg(x, c("en", "fr", "es"))
-    if (out == "en") 
+    if (out == "en")
         out <- NULL else out <- paste0("language=", out)
     out
 }
 
 rcites_scope <- function(x) {
     out <- match.arg(x, c("current", "historic", "all"))
-    if (out == "current") 
+    if (out == "current")
         out <- NULL else out <- paste0("scope=", out)
     out
 }
@@ -87,11 +90,11 @@ rcites_combine_lists <- function(x, taxon_id, raw) {
         if (!all(unlist(lapply(ls_keys, identical, ls_keys[[1L]])))) {
             stop("Cannot combine lists with different names")
         }
-        # 
+        #
         for (i in seq_along(x)) {
             x[[i]] <- lapply(x[[i]], rcites_add_taxon_id, taxon_id[wch][i])
         }
-        # 
+        #
         out <- list()
         for (i in seq_along(ls_keys[[1L]])) {
             key <- ls_keys[[1L]][i]
@@ -125,28 +128,28 @@ rcites_forgetsecret <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
 
 ##################### Pagination helpers
 
-rcites_autopagination <- function(q_url, per_page, pages, tot_page, token, 
+rcites_autopagination <- function(q_url, per_page, pages, tot_page, token,
     verbose = TRUE) {
     out <- list()
-    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$", 
+    q_url_0 <- gsub(q_url, pattern = "page=[[:digit:]]+\\&per_page=[[:digit:]]+$",
         replacement = "")
-    # 
+    #
     for (i in seq_along(pages)) {
-        if (verbose) 
-            cat("Retrieving info from page ", pages[i], "/", tot_page, 
+        if (verbose)
+            cat("Retrieving info from page ", pages[i], "/", tot_page,
                 "\r")
-        q_url_new <- paste0(q_url_0, "page=", pages[i], "&per_page=", min(per_page, 
+        q_url_new <- paste0(q_url_0, "page=", pages[i], "&per_page=", min(per_page,
             500))
         out[[i]] <- rcites_res(q_url_new, token)
     }
-    if (verbose) 
+    if (verbose)
         cat("\nDone!\n")
-    # 
+    #
     out
 }
 
 rcites_numberpages <- function(x) {
-    x$total_entries%/%x$per_page + (x$total_entries%%x$per_page > 0)
+    x$total_entries %/% x$per_page + (x$total_entries %% x$per_page > 0)
 }
 
 
@@ -161,7 +164,7 @@ rcites_add_author_year <- function(x) {
     x
 }
 
-# convert null to na
+# convert null to na recursively
 rcites_null_to_na <- function(x) {
     if (is.list(x)) {
         return(lapply(x, rcites_null_to_na))
@@ -195,14 +198,14 @@ rcites_assign_class <- function(x) {
 rcites_simplify_listings <- function(x) {
     # fields below may or may not be included, so there are removed
     vc_rm <- c("party", "hash_annotation", "annotation")
[…]
