+++ tests/testthat/helper-citesr.R
@@ -12,4 +12,4 @@ cl_df <- c("tbl_df", "tbl",  "data.frame")
 cl_raw <- c("list", "spp_raw")
 cl_raw_multi <- c("list", "spp_raw_multi")
 
-ut_pause <- function(x = 4) Sys.sleep(x)
+ut_pause <- function(x = 5) Sys.sleep(x)
