+++ tests/test-all.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(rcites)
-
-test_check("rcites")
