+++ inst/doc/get_started.R
@@ -1,9 +1,0 @@
-## ----setup, echo = FALSE-------------------------------------------------
-NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
-knitr::opts_chunk$set(
-  purl = NOT_CRAN,
-  eval = NOT_CRAN,
-  fig.align = "center",
-  comment = "#> "
-)
-
