### 2018-10-08 29457d9
review get_started vignette
