+++ inst/doc/elephant.R
@@ -1,0 +1,11 @@
+## ----setup, echo = FALSE-------------------------------------------------
+NOT_CRAN <- FALSE
+# NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
+knitr::opts_chunk$set(
+  purl = NOT_CRAN,
+  eval = NOT_CRAN,
+  fig.align = "center",
+  comment = "#> "
+)
+library(rcites)
+
