+++ vignettes/get_started.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Get started with rcites"
 author: "rcites team"
-date: 07-18-2018
+date: 10-08-2018
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Get started with rcites}
@@ -10,6 +10,7 @@ vignette: >
 
 
 ```{r setup, echo = FALSE}
+# NOT_CRAN <- FALSE
 NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
 knitr::opts_chunk$set(
   purl = NOT_CRAN,
@@ -21,15 +22,16 @@ knitr::opts_chunk$set(
 
 
 
-# Set up a connection to the Species+ database
+# Set up a connection to the Species+/CITES Checklist API
 
 ## Get your personal token
-To set up a connection to the CITES Species+ database, an authentication token
+
+To set up a connection to the Species+/CITES Checklist API, an authentication token
 is required. Each user should obtain his or her own personal token
-to run the code below (see https://api.speciesplus.net/documentation for more details).
-To obtain a token, sign up on the [Speciesplus API website](http://api.speciesplus.net/users/sign_up).
+to run the code below (see https://api.speciesplus.net/documentation for more details). To obtain a token, sign up on the [Species+ API website](http://api.speciesplus.net/users/sign_up).
 
 ## Set the token
+
 Now, we assume that you already have a token. For illustrative purposes,
 we will use the generic token value `8QW6Qgh57sBG2k0gtt` for the API
 documentation.
@@ -41,59 +43,60 @@ requests. They are three different ways to set the token in `rcites`:
 [`.Renviron`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/Startup.html)
 file (preferred);
 
-2. use `set_token()` to set up the token as character string (with quotes). If not
-entered in the function, the token can be passed without quotes (not as character
-string) after a prompt. This way, the token `SPECIESPLUS_TOKEN` is interactively
-set up for the current R session (meaning a login will be required for the future
-R sessions);
+2. use `set_token()` to set up the token as character string (with quotes). If not entered in the function, the token can be passed without quotes (not as character string) after a prompt. This way, the token `SPECIESPLUS_TOKEN` is interactively set up for the current R session (meaning a login will be required for the future R sessions);
 
 ```{r, eval = FALSE}
 set_token("8QW6Qgh57sBG2k0gtt")
 ```
 
-3. use the `token` argument of the functions, *i.e.* the token is passed manually
-to each function call.
+3. use the `token` argument of the functions, *i.e.* the token is passed manually to each function call.
 
 
 # Use of key features
 
-## Access the Speciesplus taxon concept and retrieve a taxon id
+## Retrieve a taxon identifiers with spp_taxonconcept()
 
-To efficiently query information from the CITES Speciesplus database, you
-first need to retrieve the unique taxon identifier `taxon_id` from the [Speciesplus taxon concept](https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html).
+### Basic calls to spp_taxonconcept()
+
+In order to efficiently query information from the CITES database, you
+first need to retrieve the unique taxon identifier `taxon_id` from the [Species+ taxon concept](https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html).
 To do so, you should first call `spp_taxonconcept()` and provide the scientific
[…]

[…]
