### 2018-10-09 cbbefe3
re-work vignette get_started :pencil:
