+++ R/set_token.R
@@ -1,12 +1,13 @@
 #' Login helper function
 #'
-#' This function sets the authentification token for the current session.
+#' Set and forget the authentification token for the current session.
 #'
 #' @param token a character string (with quotes) containing you token. If `NULL` then the
 #' token can be passed without quotes (not as character string) after a prompt.
 #'
 #' @references
 #' \url{https://api.speciesplus.net/documentation}
+#'
 #'
 #' @export
 #'
@@ -18,8 +19,9 @@
 #'  set_token()
 #' }
 
+#' @describeIn set_token set the environment variable `SPECIESPLUS_TOKEN`.
 set_token <- function(token = NULL) {
-    if (is.null(token)) 
+    if (is.null(token))
         token <- readline("Enter your token without quotes: ")
     if (identical(token, "")) {
         message("no token provided")
@@ -29,3 +31,7 @@ set_token <- function(token = NULL) {
     }
     invisible(NULL)
 }
+
+#' @describeIn set_token forget the environment variable `SPECIESPLUS_TOKEN`.
+#' @export
+forget_token <- function() Sys.unsetenv("SPECIESPLUS_TOKEN")
