+++ R/print.spp.R
@@ -1,27 +1,26 @@
-#' Print methods for objects of class `spp_*`
+#' Print methods for objects of class `spp_raw*`
 #'
 #' Print the outputs of a Species+ API call
 #'
-#' @param x The result object.
-#' @param ... Ignored.
-#'
-#' @name print.spp
+#' @param x An object of class `spp_raw*`.
+#' @param ... ignored.
 #'
 #' @return The JSON result.
 #'
-#' @importFrom jsonlite prettify toJSON
-NULL
+#' @name print.spp
 
+NULL
 
 #' @rdname print.spp
 #' @method print spp_raw
+#' @export
 print.spp_raw <- function(x, ...) {
-    print(toJSON(unclass(x), pretty = TRUE, auto_unbox = TRUE))
+    print(jsonlite::toJSON(unclass(x), pretty = TRUE, auto_unbox = TRUE))
 }
-
 
 #' @rdname print.spp
 #' @method print spp_raw_multi
+#' @export
 print.spp_raw_multi <- function(x, ...) {
     print.spp_raw(x)
     cat("\n")
@@ -30,9 +29,102 @@ print.spp_raw_multi <- function(x, ...) {
 }
 
 
-
+#' @method print spp_cites_leg
 #' @rdname print.spp
+#' @export
+print.spp_cites_leg <- function(x, ...) {
+
+    rcites_print_title("Cites listings ($cites_listings):", "\n")
+    rcites_print_df_rm(x$cites_listings, col_rm = c("annotation", "hash_annotation"))
+
+    rcites_print_title("Cites quotas ($cites_quotas):", "\n", "\n")
+    rcites_print_df_rm(x$cites_quotas, col_rm = c("notes", "url"))
+
+    rcites_print_title("Cites suspensions ($cites_suspensions):", "\n",
+        "\n")
+    rcites_print_df_rm(x$cites_suspensions, col_rm = c("notes", paste0("start_notification.",
+        c("name", "date", "url"))))
+}
+
+#' @method print spp_cites_leg_multi
+#' @rdname print.spp
+#' @export
+print.spp_cites_leg_multi <- function(x, ...) {
+    rcites_print_taxon_id(x$cites_listings$taxon_id)
+    cat("\n")
+    print.spp_cites_leg(x)
+}
+
+
+#' @method print spp_distr
+#' @rdname print.spp
+#' @export
+print.spp_distr <- function(x, ...) {
+    rcites_print_title("Distributions ($distributions):", "\n")
+    rcites_print_df(x$distributions)
+
+    rcites_print_title("References ($references):", "\n", "\n")
+    x$references$reference <- rcites_print_shorten(x$references$reference)
+    rcites_print_df_rm(x$references)
+}
+
+#' @method print spp_distr_multi
+#' @rdname print.spp
+#' @export
+print.spp_distr_multi <- function(x, ...) {
+    rcites_print_taxon_id(x$distributions$taxon_id)
+    cat("\n")
+    print.spp_distr(x)
+}
+
+
+#' @method print spp_eu_leg
+#' @rdname print.spp
+#' @export
+print.spp_eu_leg <- function(x, ...) {
+
+    rcites_print_title("EU listings ($eu_listings):", "\n")
+    rcites_print_df_rm(x$eu_listings, col_rm = c("annotation", "hash_annotation"))
+
+    rcites_print_title("EU decisions ($eu_decisions):", "\n", "\n")
+    npr <- c("notes", "eu_decision_type.description", "start_event.url")
+    rcites_print_df_rm(x$eu_decisions, col_rm = npr)
+}
+
+#' @method print spp_eu_leg_multi
+#' @rdname print.spp
+#' @export
[…]
