### 2018-10-09 ec0a9d4
draft bulk_analysis
