+++ inst/doc/elephant.Rmd
@@ -10,8 +10,8 @@ vignette: >
 
 
 ```{r setup, echo = FALSE}
-# NOT_CRAN <- FALSE
-NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
+NOT_CRAN <- FALSE
+# NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
 knitr::opts_chunk$set(
   purl = NOT_CRAN,
   eval = NOT_CRAN,
