+++ vignettes/bulk_analysis.Rmd
@@ -1,0 +1,59 @@
+---
+title: "Bulk analysis with rcites"
+author: "rcites team"
+date: 10-08-2018
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Bulk analysis with rcites}
+  %\VignetteEngine{knitr::rmarkdown}
+---
+
+
+```{r setup, echo = FALSE}
+NOT_CRAN <- FALSE
+# NOT_CRAN <- identical(tolower(Sys.getenv("NOT_CRAN")), "true")
+knitr::opts_chunk$set(
+  purl = NOT_CRAN,
+  eval = NOT_CRAN,
+  fig.align = "center",
+  comment = "#> "
+)
+library(rcites)
+```
+
+## `spp_distributions()` for broad queries
+
+`spp_distributions()` 
+
+```{r}
+res_cms <- spp_taxonconcept(".", taxonomy = "CMS")
+dim(res_cms$general)
+```
+
+
+```{r}
+res_cites <- spp_taxonconcept(".", page = 1:3)
+dim(res_cms$general)
+```
+
+
+## Retrieving information for a set of taxon_concept
+
+`spp_distributions()`, `spp_eu_legislation()`, `spp_cites_legislation()` and
+`spp_references()` handle vector of taxon_id which allow of bulk analysis.
+
+```{r}
+res1 <- spp_distributions(taxon_id = c('4521', '3210', '10255'))
+```
+
+```{r}
+res2 <- spp_cites_legislation(taxon_id = c('4521', '3210', '10255'))
+```
+
+```{r}
+res3 <- spp_eu_legislation(taxon_id = c('4521', '3210', '10255'))
+```
+
+```{r}
+res4 <- spp_references(taxon_id = c('4521', '3210', '10255'))
+```
