### 2018-10-09 aa1d319
fix link in Readme + add link in vignette
