+++ vignettes/get_started.Rmd
@@ -179,3 +179,8 @@ spp_distributions("4521", verbose = FALSE)
 ```{r}
 spp_references("4521", verbose = FALSE)
 ```
+
+
+## Bulk analysis
+
+Please see [Bulk analysis with rcites](https://ibartomeus.github.io/rcites/articles/bulk_analysis.html) for details.
