### 2018-10-09 b552ef9
add back test-all.R
