+++ tests/test-all.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(rcites)
+
+test_check("rcites")
