### 2018-10-09 819237d
coerce taxon_id to character when outputs are combined
