### 2018-10-09 3dbd8dc
first output tests :shield:
