+++ .travis.yml
@@ -33,3 +33,4 @@ r_github_packages:
 
 after_success:
   - Rscript -e 'covr::codecov()'
+  - Rscript -e 'covr::coveralls()'
