### 2018-10-12 af8eca9
edit acknowledgements and citation
