### 2018-10-12 5d56bd9
vignette cleanup
