### 2018-10-12 1c6c16d
add vignettes to readme
