### 2018-10-12 d684133
minor rewording
