### 2018-10-12 765fac1
explain vignettes
