### 2018-10-15 7ef4927
Correct a few typos and small clarifications.
