+++ vignettes/elephant.Rmd
@@ -21,10 +21,6 @@ knitr::opts_chunk$set(
 library(rcites)
 ```
 
-
-<!-- https://cran.r-project.org/web/packages/httr/vignettes/secrets.html -->
-
-
 # Introduction and setup
 
 In [Get started with rcites](https://ibartomeus.github.io/rcites/articles/get_started.html), we explained how to get a token and set it up for general access to the CITES Species+ database. Also, we very briefly introduced to how to code the key features of `rcites`. With this article, we aim to further introduce to the functionality and workflows of `rcites`. For this, we use the African bush elephant (*Loxodonta africana*, hereafter "elephant") as a case study. The elephant not only is a highly endangered species that is illegally traded globally but also a flagship species of nature conservation and the logo species of CITES.
@@ -35,7 +31,6 @@ We start with a basic set up: we load the package and set the token:
 library(rcites)
 set_token("8QW6Qgh57sBG2k0gtt")
 ```
-
 
 # Retrieve the taxon id
 
@@ -52,7 +47,7 @@ elephant_taxonconcept
 
 As the first column of the output shows, the taxon identifier of the elephant is 4521. This `taxon_id` will be used for all next function coding.
 
-Beyong the taxon identifier, the output also provides information about the taxon classification and other names, both synonyms and common names if any, in different languages.
+Beyond the taxon identifier, the output also provides information about the taxon classification and other names, both synonyms and common names if any, in different languages.
 
 
 ## Map the elephant's distribution
@@ -100,12 +95,12 @@ elephant_eu
 ```
 
 
-With the combination of map2 and the legislation data, one might be able to illustrate the elephant's trade directions. This and other use examples of the `rcites` data output will be added bit by bit.
+With the combination of `map2` and the legislation data, one might be able to illustrate the elephant's trade directions. This and other use examples of the `rcites` data output will be added bit by bit.
 
 
 ## Access the elephant's Species+ reference data
 
-Last but not least, especially scientists will probably be interested in what references the Species+ information about the elephant is based on. For this, we can access the Species+ reference data with the `spp_references()` function.
+Last but not least, it is important to identify which references the Species+ information about the elephant is based on. For this, we can access the Species+ reference data with the `spp_references()` function.
 
 ```{r}
 elephant_refs <- spp_references(taxon_id = "4521")
