+++ vignettes/get_started.Rmd
@@ -33,31 +33,26 @@ to run the code below (see https://api.speciesplus.net/documentation for more de
 ## Set the token
 
 Now, we assume that you already have a token. For illustrative purposes,
-we will use the generic token value `8QW6Qgh57sBG2k0gtt` for the API
+we will use the generic token value `8QW6Qgh57sBG2k0gtt` from the API
 documentation.
 
 A token is mandatory and needs to be passed to the header of all URL
 requests. They are three different ways to set the token in `rcites`:
 
-1. set an environment variable `SPECIESPLUS_TOKEN` in
-[`.Renviron`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/Startup.html)
-file (preferred);
+1. set an environment variable `SPECIESPLUS_TOKEN` in your [`.Renviron`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/Startup.html)
+file (preferred for frequent users);
 
-2. use `set_token()` to set up the token as character string (with quotes). If not entered in the function, the token can be passed without quotes (not as character string) after a prompt. This way, the token `SPECIESPLUS_TOKEN` is interactively set up for the current R session (meaning a login will be required for the future R sessions);
+2. use `set_token()` to set up the token as a character string (with quotes). If not entered in the function, the token can be passed without quotes (not as a character string) after the prompt. This way, the token `SPECIESPLUS_TOKEN` is interactively set up only for the current R session, meaning a login will be required for the future R sessions (preferred);
 
 ```{r, eval = FALSE}
 library(rcites)
 set_token("8QW6Qgh57sBG2k0gtt")
 ```
 
-```{r, echo = FALSE}
-load_all()
-```
+3. use the `token` argument inside the functions, *i.e.* the token is passed manually to each function call.
 
-3. use the `token` argument of the functions, *i.e.* the token is passed manually to each function call.
-
-Note that if set the token wrongly and wish for setting it again interactively,
-you must first forget the previous one with `forget_token()`.
+Note that if you set a wrong token and you wish to set it again interactively,
+you must first forget the previous token with `forget_token()`.
 
 # Use of key features
 
@@ -68,13 +63,13 @@ you must first forget the previous one with `forget_token()`.
 In order to efficiently query information from the CITES database, you
 first need to retrieve the unique taxon identifier `taxon_id` from the [Species+ taxon concept](https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html).
 To do so, you should first call `spp_taxonconcept()` and provide the scientific
-name of the taxon you look for. Let us start by requesting the identifier of the African bush elephant, *i.e.* *Loxodonta africana*.
+name of the taxon you are looking for. Let us start by requesting the identifier of the African bush elephant, *i.e.* *Loxodonta africana*.
 
 ```{r result_1}
 res1 <- spp_taxonconcept(query_taxon = "Loxodonta africana")
 ```
 
-Note that if you chose chose option 3, then the code should look like the one below:
+Note that if you have decide to set your token using the third option, then the code should look like the one below:
 
 ```{r, eval = FALSE}
 res1 <- spp_taxonconcept(query_taxon = "Loxodonta africana", token = "8QW6Qgh57sBG2k0gtt")
@@ -96,11 +91,11 @@ res1
 ```
 
 For some taxa, there are more than one taxon identifier available. In `general` only
-*active* identifiers are listed, the the full list of identifier is
[…]

[…]
