+++ R/set_token.R
@@ -2,8 +2,9 @@
 #'
 #' Set and forget the authentification token for the current session.
 #'
-#' @param token a character string (with quotes) containing you token. If `NULL` then the
-#' token can be passed without quotes (not as character string) after a prompt.
+#' @param token a character string (with quotes) containing you token. If `NULL`
+#' then the token can be passed without quotes (not as character string) after
+#' a prompt.
 #'
 #' @references
 #' \url{https://api.speciesplus.net/documentation}
