+++ man/set_token.Rd
@@ -10,8 +10,9 @@ set_token(token = NULL)
 forget_token()
 }
 \arguments{
-\item{token}{a character string (with quotes) containing you token. If \code{NULL} then the
-token can be passed without quotes (not as character string) after a prompt.}
+\item{token}{a character string (with quotes) containing you token. If \code{NULL}
+then the token can be passed without quotes (not as character string) after
+a prompt.}
 }
 \description{
 Set and forget the authentification token for the current session.
