### 2018-10-24 #244 @noamross
The authors have done a fantastic job with revisions.  I must say that the new iteration of the package is a pleasure to use - the  set of functions are simplified and intuitive, the print methods are clear and useful, the processed data is easy to use, and bulk analysis is straightforward.  Even the internal code structure has improved and is easy to follow. I heartily recommend acceptance!  I have checked off the boxes above.

One small note: I see that when the results from vectorized calls are combined into aggregated data frames, the tibble classes are dropped.  I think this should be a quick fix to `rcites_combine_lists()`
