### 2018-10-24 #244 @KevCaz
Thanks @noamross, you and @mcsiple have been **very** helpful in this process.
Just a quick question/confirmation, are you Ok to be added as a reviewer in the DESCRIPTION file?

Regarding:

> One small note: I see that when the results from vectorized calls are combined into aggregated data frames, the tibble classes are dropped. I think this should be a quick fix to rcites_combine_lists()

We'll take care of it asap :)
