### 2018-10-24 ibartomeus/rcites#38 @KevCaz
rcites_combine_lists() dropped tibble classes 
As mentioned by Noam [here](https://github.com/ropensci/onboarding/issues/244#issuecomment-432821342):

> I see that when the results from vectorized calls are combined into aggregated data frames, the tibble classes are dropped. I think this should be a quick fix to rcites_combine_lists()
