### 2018-10-25 0ce8c62
rcites_combine_lists() do not drop tibble classes anymore close #38
