+++ inst/assets/draft_answer.md
@@ -1,19 +1,13 @@
-We are grateful to the reviewers for the time they spent on reviewing our package and for the quality of their comments that significantly contributed to
-improve our package. We added both reviewers on the list of contributors
-in `DESCRIPTION` with the role "rev" and we would like to have their formal
-consent before the next CRAN release.
+We are extremely grateful to the reviewers (@noamross @mcsiple) for the time they spent on reviewing our package and for the quality of their comments that significantly contributed to improve `rcites`. We added both reviewers on the list of contributors in `DESCRIPTION` with the role "rev" and we would like to have their formal consent before the next CRAN release.
 
-We have carefully addressed the reviewers' comments and added several
-features to our package following their suggestions. Given the nature and the
-number of changes we have made since the last version, we are considering
-the new version as a major release (the future CRAN release would be v1.0.0).
+We have carefully addressed the reviewers' comments and added several features to our package following their suggestions. Given the nature and the number of changes we have made since the last version, we are considering the new version as a major release (the future CRAN release would be v1.0.0). Below we detail these changes.
+
 
 
 
 ## Changes in functions
 
-
-### data.table is no longer a dependency
+### `data.table` is no longer a dependency
 
 We agree with Noam Ross that the use of `data.table` was not particularly relevant for our package we therefore removed the dependencies to this package. This was handled in [#34](https://github.com/ibartomeus/rcites/pull/34).
 
@@ -25,41 +19,39 @@ As suggested by Noam Ross, all outputs are now S3 objects with specific print me
 
 ### New parameters to better include the API features
 
-Our functions now integrate all parameters of the Species+ API, meaning:
+Our functions now integrate all parameters of the Species+ API, that is:
 
-- `taxonomy` (the package now handles requests for the "CMS" data base), `with_descendants`, `language`, `updated_since`, `per_page`, `pages` (that enables to select any a set of pages) were added to `spp_taxonconcept()`,
+- `taxonomy` (NB the package now handles requests for the "CMS" data base), `with_descendants`, `language`, `updated_since`, `per_page`, `pages` (that enables to select any a set of pages) were added to `spp_taxonconcept()`,
 - `scope` and `language` were added to `spp_eu_legislation()` and `spp_cites_legislation()`,
 - `language` was added to `spp_distributions()`
 
 ### Bulk analysis
 
-In order to allow the user to perform bulk analysis using `spp_taxonconcept()`,
-we added an auto-pagination to the  `spp_taxonconcept()`. The user is now
-able to retrieve more than 500 entries at once (*i.e.* all entries of the data base when `query_taxon=""`).
-When more than one page must be fetched, a message is prompted (if `verbose` is set to `TRUE`).
-All these changes required a new set of tests handled in [#35](https://github.com/ibartomeus/rcites/pull/35).
+In order to allow the user to perform bulk analyses with `spp_taxonconcept()`,
+we added an auto-pagination inside `spp_taxonconcept()`. The user is now
+able to retrieve more than 500 entries at once (*i.e.* all entries of the data base when `query_taxon=""`). When more than one page must be fetched, a message is prompted (if `verbose` is set to `TRUE`). All these changes required a new set of tests handled in [#35](https://github.com/ibartomeus/rcites/pull/35).
 
 The four remaining `spp_*` functions now support `taxon_id` with more than one
-element. This was handled in [#36](https://github.com/ibartomeus/rcites/pull/36),
-in this case we decided to use a different classes when the output is a combination
-of outputs of different species. Basically, we concatenated the lists and add a column
-`taxon_id` to all data frames. Such S3 object have class `*_multi` where `*` stands
-for the name of the class for a single element.
[…]
