### 2018-10-25 #244 @JonasGeschke
Thank you @noamross for your kind words! We did and will continue to do our best to improve the package, thank you again for your comprehensive review.
