### 2018-11-12 ibartomeus/rcites#39 @KevCaz
Editor's comments on version 0.1.0.9000 
- [x] I see you have examples wrapped in \donttest{}, AFAIK these examples are run on cran's checks, is that what you intended? if not, change it to \dontrun{}
 - [x] In running goodpractice::gp(), i see there's quite a few files that still have lines longer than 80 width, run gp, then see x$lintr - within reason try to make these width 80 or less
 - [x] I'd consider adding a parameter ... to all functions that do HTTP requests to allow users to pass curl options to any HTTP request. I personally consider it best practice, though it's not in our policies. I guess an alternative is that users can just wrap your functions in with_verbose(...) if they want to use curl options. this could go under something to think about moving forward (blog post https://ropensci.org/blog/2014/12/18/curl-options/ covering curl options)

See https://github.com/ropensci/onboarding/issues/244
