### 2018-11-12 #244 @KevCaz
Thanks @sckott, your help is very much appreciated.

`\donttest{}` was a recommendation by one of the CRAN's maintainers:

> Thanks, please do not comment out your examples and use \donttest{} instead:
>\examples{
   examples for users and checks:
   executable in < 5 sec
 >  donttest{
       further examples for users (not used for checks)
   }
}

After using  `\donttest{}`, we did not have any issue with our CRAN submission. I'll have a look at the difference between the two.
