### 2018-11-12 #244 @sckott
interesting. I guess donttest examples aren't run on cran's check now.
