### 2018-11-12 40b87eb
add ... to HTTP request see #39
