+++ R/spp_eu_legislation.R
@@ -18,6 +18,7 @@
 #' set directly in `Renviron`. Alternatively, [set_token()] can
 #' be used to set `SPECIESPLUS_TOKEN` for the current session.
 #' @param verbose a logical. Should extra information be reported on progress?
+#' @param ... Further named parameters, see [httr::GET()].
 #'
 #' @return If `raw` is set to `TRUE` then an object of class `spp_raw` (or
 #' `spp_raw_multi` if `length(taxon_id)>1`) is returned which is essentially
@@ -37,14 +38,14 @@
 #' res1 <- spp_eu_legislation(taxon_id = '4521')
 #' res2 <- spp_eu_legislation(taxon_id = c('4521', '3210', '10255'))
 #' res3 <- spp_eu_legislation(taxon_id = '4521', scope = 'historic')
-#' res4 <- spp_eu_legislation(taxon_id = '4521', scope = 'all', language='fr')
+#' res4 <- spp_eu_legislation(taxon_id = '4521', scope = 'all', language='fr', config=httr::verbose())
 #' }
 
 spp_eu_legislation <- function(taxon_id, scope = "current", language = "en",
-    raw = FALSE, token = NULL, verbose = TRUE) {
+    raw = FALSE, token = NULL, verbose = TRUE, ...) {
     if (length(taxon_id) > 1) {
         out <- lapply(taxon_id, spp_eu_legislation, scope = scope,
-          language = language, raw = raw, token = token, verbose = verbose)
+          language = language, raw = raw, token = token, verbose = verbose, ...)
         out <- rcites_combine_lists(out, taxon_id, raw)
     } else {
         # token check
@@ -65,7 +66,7 @@ spp_eu_legislation <- function(taxon_id, scope = "current", language = "en",
             q_url <- rcites_url("taxon_concepts/", taxon_id,
               "/eu_legislation.json", query_string)
             ## get_res
-            tmp <- rcites_res(q_url, token)
+            tmp <- rcites_res(q_url, token, ...)
             ## outputs
             if (raw) {
                 out <- tmp
