### 2018-11-12 c7addb2
shorten lines
