+++ tests/testthat/test-wrongtoken.R
@@ -12,5 +12,6 @@ test_that("expected errors", {
   ut_pause()
   expect_error(rcites_res(q_url, token), msg, fixed = TRUE)
   ut_pause()
-  expect_error(spp_taxonconcept(query_taxon = tx_nm, token = token), msg, fixed = TRUE)
+  expect_error(spp_taxonconcept(query_taxon = tx_nm, token = token), msg,
+    fixed = TRUE)
 })
