### 2018-11-12 ropensci/rcites#39 @KevCaz
Done via 7f17d56972d7ca900627c55f7b6e668b35c83e7b
