### 2018-11-12 #244 @KevCaz
@sckott  I follow your recommendations, see https://github.com/ibartomeus/rcites/issues/39.
