### 2018-11-12 #244 @sckott
you mean test it in units tests?
