### 2018-11-13 #244 @sckott
@KevCaz sometimes I forget, but yes, e.g., you could test that timeout works, e.g.,

```r
expect_error(some_fxn(httr::timeout(10)), "some message")
# or just that it errors with any message
expect_error(some_fxn(httr::timeout(10)))
```
