+++ tests/testthat/test-legislations.R
@@ -3,7 +3,6 @@ context("Legislations")
 skip_on_cran()
 skip_if_no_auth()
 ut_pause()
-
 #
 res1 <- spp_cites_legislation(taxon_id = tx_id, verbose = FALSE)
 res1b <- spp_cites_legislation(taxon_id = tx_id2, verbose = FALSE)
