+++ cran-comments.md
@@ -1,26 +1,22 @@
-## This is our first submission
+## This is our second submission
 
-NB: Following the first review, I now use \donttest{} instead of commenting out
-examples.
 
 ## Test environments
 
 * local OS X install (10.13.6), R 3.5.1
-* local debian testing (4.16.0-2-amd64), R 3.5.1
-* ubuntu 14.04 (on travis-ci), R 3.4.4, 3.5.0, R-devel
-* win-builder (devel and release)
+* local Debian testing (4.18.0-2-amd64), R 3.5.1
+* Ubuntu 14.04 (on Travis-ci), R 3.4.4, R:release, R:devel
+* win-builder (R:release and R:devel)
 
 
 ## R CMD check results
 
-There were no ERRORs or WARNINGs.
+There were no ERRORs, WARNINGs or NOTEs.
 
-1 NOTE, New submission:
-
-* checking CRAN incoming feasibility ... NOTE
-Maintainer: 'Kevin Cazelles <kevin.cazelles@gmail.com>'
-
-New submission
+Installation time in seconds: 7
+Check time in seconds: 99
+Status: OK
+R Under development (unstable) (2018-11-12 r75592)
 
 
 ## Downstream dependencies
