### 2018-11-14 9788608
add ROpenSci badge and citation edit
