### 2018-11-15 #244 @ibartomeus
Great! I can draft a blog post (and let @KevCaz and @JonasGeschke add stuff before doing anything else). @stefaniebutland let me know how to proceed.
