### 2018-11-16 7c38a84
make the repo ready for the release of v1.0.0
