+++ README.md
@@ -3,9 +3,8 @@
 ### Current Status
 
 [![Build status](https://travis-ci.org/ropensci/rcites.svg?branch=master)](https://travis-ci.org/ropensci/rcites)
-[![Build status](https://ci.appveyor.com/api/projects/status/bsd3c7mv12xv959j/branch/master?svg=true)](https://ci.appveyor.com/project/KevCaz/rcites/branch/master)
+[![Build status](https://ci.appveyor.com/api/projects/status/kgimo4v7rvtpkp5e?svg=true)](https://ci.appveyor.com/project/KevCaz/rcites-mo3vb)
 [![codecov](https://codecov.io/gh/ropensci/rcites/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/rcites)
-[![Coverage Status](https://coveralls.io/repos/github/ropensci/rcites/badge.svg?branch=master)](https://coveralls.io/github/ropensci/rcites?branch=master)
 [![ROpenSci status](https://badges.ropensci.org/244_status.svg)](https://github.com/ropensci/onboarding/issues/244)
 [![CRAN status](https://www.r-pkg.org/badges/version/rcites)](https://www.r-pkg.org/badges/version/rcites)
 [![CRAN downloads](https://cranlogs.r-pkg.org/badges/grand-total/rcites)](https://cran.r-project.org/package=rcites)
@@ -13,7 +12,7 @@
 
 # rcites
 
-An R package to access information from the [Speciesplus](https://speciesplus.net/) database via the [Species+/CITES Checklist API](https://api.speciesplus.net/documentation/v1.html). The package is available for download from [CRAN](https://cran.r-project.org/package=rcites) (stable version) and [Github](https://github.com/ibartomeus/rcites) (development version).
+An R package to access information from the [Speciesplus](https://speciesplus.net/) database via the [Species+/CITES Checklist API](https://api.speciesplus.net/documentation/v1.html). The package is available for download from [CRAN](https://cran.r-project.org/package=rcites) (stable version) and [Github](https://github.com/ropensci/rcites) (development version).
 
 Please see the [release paper](link to come) for background information about the Convention on International Trade in Endangered Species of Wild Fauna and Flora ([CITES](https://cites.org)), the Species+ database and basic information about the aim of the package.
 
@@ -30,7 +29,7 @@ library("rcites")
 The development version can be installed via the `devtools` package:
 
 ```R
-devtools::install_github("ibartomeus/rcites")
+devtools::install_github("ropensci/rcites")
 library("rcites")
 ```
 
@@ -39,12 +38,12 @@ library("rcites")
 
 To set up a connection to the CITES Species+ database, a personal authentication
 token is required. Please see the vignette for details how to get a token and
-how to set the token for package use: [Get started with rcites](https://ibartomeus.github.io/rcites/articles/get_started.html)
+how to set the token for package use: [Get started with rcites](https://ropensci.github.io/rcites/articles/get_started.html)
 
 Additional information about specific use examples are provided for the
-[African bush elephant (*Loxodonta africana*)](https://ibartomeus.github.io/rcites/articles/elephant.html).
+[African bush elephant (*Loxodonta africana*)](https://ropensci.github.io/rcites/articles/elephant.html).
 The package usage for querying multiple species is described in another
-vignette entitled ['Bulk analysis with rcites'](https://ibartomeus.github.io/rcites/articles/bulk_analysis.html).
+vignette entitled ['Bulk analysis with rcites'](https://ropensci.github.io/rcites/articles/bulk_analysis.html).
 
 
 ### Key features
@@ -69,13 +68,13 @@ The package functions have three different prefixes:
 
 ### Citation information
 
[…]
