+++ .github/issue_template.md
@@ -1,4 +1,8 @@
-<!-- If this issue relates to usage of the package, whether a question, bug or similar, along with your query, please paste your devtools::session_info() or sessionInfo() into the code block below. If not, delete all this and proceed :) -->
+<!--
+If this issue relates to usage of the package, whether a question, bug or similar,
+along with your query, please paste your devtools::session_info() or sessionInfo()
+into the code block below. If not, delete all this and proceed :)
+-->
 
 <details> <summary><strong>Session Info</strong></summary>
 
