+++ _pkgdown.yml
@@ -7,3 +7,5 @@ authors:
     href: https://kevcaz.github.io/
   Ignasi Bartomeus:
     href: https://bartomeuslab.com/
+  Jonas Geschke:
+    href: https://www.linkedin.com/in/jonasgeschke/?originalSubdomain=de

[…]
