### 2018-11-16 ecd1be8
make the repo ready for the release of v1.0.0
