### 2018-11-16 #244 @KevCaz
@sckott one more question, what kind of access to the repo do we have? Are we allowed to change the settings of the repo? If yes I cannot do so currently! If no, can you just edit the URL at the home page of the GH repo: https://ibartomeus.github.io/rcites/ => https://ropensci.github.io/rcites/. Thanks!
