### 2018-11-16 #244 @sckott
these 3

now have admin access - now you should be able to change those things - of course you can manage access of your other contibutors
