### 2018-11-16 #244 @stefaniebutland
@JonasGeschke Great to hear you'd like to contribute a post!

[This link](https://ropensci.org/tags/onboarding/) will give you many examples of blog posts by authors of onboarded packages so you can get an idea of the style and length you prefer.

Here are some technical and editorial guidelines for contributing a blog post: https://github.com/ropensci/roweb2#contributing-a-blog-post. We ask that you submit your draft post via pull request a week before the planned publication date so we can give you some feedback.
I have 2018-12-18 available for publication. Would submission by 2018-12-11 work for you?

Happy to answer any questions.
