### 2018-11-16 #244 @JonasGeschke
@stefaniebutland thanks for your message, we are very happy to contribute a blog post. @ibartomeus I just put you in the loop so you also see the above two links.
submission by 2018-12-11 should be doable @ibartomeus ?!
