+++ cran-comments.md
@@ -1,15 +1,14 @@
 ## This is our second submission
 
-
 ## Test environments
 
-* local OS X install (10.13.6), R 3.5.1
 * local Debian testing (4.18.0-2-amd64), R 3.5.1
+* MacOSX 10.13.3 (on Travis-ci), R:release
 * Ubuntu 14.04 (on Travis-ci), R 3.4.4, R:release, R:devel
 * win-builder (R:release and R:devel)
 
 
-## R CMD check results
+## R CMD check results (win-builder)
 
 There were no ERRORs, WARNINGs or NOTEs.
 
