### 2018-11-17 360b95b
add codemeta.json + repos status badge + rebuild pkgdown website
