### 2018-11-17 #244 @KevCaz
@sckott  just to let you know that `codemeta.json`is now added.
