### 2018-11-18 #244 @JonasGeschke
> @sckott just to let you know that `codemeta.json`is now added.

and with this, all the above points (https://github.com/ropensci/onboarding/issues/244#issuecomment-438742734) are done.
anything else you need us to do for now?

Zenono DOI also is done, as well as a CRAN release 1.0.0. thus I guess we are ready for the JOSS submission, right?
(btw I dont know how long it then needs for the paper to pop up, but on wednesday there is the wildlife forum at the CBD COP-14. it would be a great opportunity to promote the paper and package.)
