### 2018-11-18 #244 @stefaniebutland
> on wednesday there is the wildlife forum at the CBD COP-14. it would be a great opportunity to promote the paper and package

@JonasGeschke Seems reasonable that I could send a tweet from rOpenSci about the package if that might help? If you think this is a good idea, please send me a short phrase or two that I can use in tweet about pkg along with twitter handle and/or relevant conference hashtag, your twitter handle & that of any pkg co-authors I should tag.
