### 2018-11-19 #244 @sckott
@JonasGeschke no, nothing else. thanks much.

* For JOSS:
    * Looks like there's already a Zenodo web hook, so that's good.
    * If you haven't already, generate a new release associated with a git tag in your repo. This doesn't mean you have to submit to CRAN with that git tag/release, its just to get a Zenodo DOI - and you can put a zenodo badge in your readme, then after each git tag it will generate a new Zenodo release
    * Then submit your `paper.md` to JOSS at http://joss.theoj.org/papers/new

I'm not familiar with the editors at JOSS - Not sure how long it takes for review at JOSS - but I guess you could just link to the paper in your repo / zenodo?
