### 2018-11-19 #244 @noamross
@JonasGeschke You can certainly talk about it as "in press" if you want.  If you want to accelerate processing, once you submit, you can go to the automatedly generate GitHub issue at https://github.com/openjournals/joss-reviews/issues, and comment that the package has been accepted here, linking back to this issue.
