### 2018-11-20 #244 @JonasGeschke
Thanks for your messages. So I submitted the paper to JOSS. I will let them know about the ropensci review process as told in https://ropensci.github.io/dev_guide/onboarding-guide-for-editors.html#for-joint-joss-submissions

> If you haven't already, generate a new release associated with a git tag in your repo. This doesn't mean you have to submit to CRAN with that git tag/release, its just to get a Zenodo DOI - and you can put a zenodo badge in your readme, then after each git tag it will generate a new Zenodo release

@KevCaz you got this, right?

@stefaniebutland thanks, a tweet would be great! I will prepare you a text!

Then I guess this issue can be closed ... but this I believe is your task @sckott . Thanks for all your efforts!
