### 2018-11-20 008dc80
minor paper changes
