+++ paper.bib
@@ -18,7 +18,7 @@ publisher = {CITES Secretariat, Geneva, Switzerland}
 
 @misc{Speciesplus_about,
 author = {{UNEP}},
-year = {n.d. a},
+year = {n.d.},
 title = {About Species+},
 doi = {},
 howpublished = {\url{https://speciesplus.net/about}},
@@ -27,7 +27,7 @@ publisher = {UN Environment, Nairobi, Kenya. Compiled by UNEP-WCMC, Cambridge, U
 
 @misc{UNEP,
 author = {{UNEP}},
-year = {n.d. b},
+year = {n.d.},
 title = {The Species+ Website},
 doi = {},
 howpublished = {\url{https://speciesplus.net}},
