### 2018-11-20 dd95790
minor paper changes
