### 2018-11-20 839a318
add launch paper
