+++ _pkgdown.yml
@@ -8,4 +8,4 @@ authors:
   Ignasi Bartomeus:
     href: https://bartomeuslab.com/
   Jonas Geschke:
-    href: https://www.linkedin.com/in/jonasgeschke/?originalSubdomain=de
+    href: https://www.linkedin.com/in/jonasgeschke/
