### 2018-11-20 #244 @JonasGeschke
So the JOSS paper is published, this is great!

@stefaniebutland here its the tweet you can do tomorrow (21.11.2018):

> It is #WildlifeForum today, promoting the sustainable use for conservation and livelihoods. #rcites helps with accessing the @CITES #Speciesplus #trade database: https://doi.org/10.21105/joss.01091
by @J_Geschke @KCazelles @ibartomeus
#EgyptCOP14 #COP14 @UNEnvironment @unepwcmc @EU_ENV
