### 2018-11-20 #244 @stefaniebutland
Thanks @JonasGeschke. I'll adapt it to fit rOpenSci tweet style and schedule it for 21.11.2018.
