### 2018-11-25 0bd91da
fix a :bug: in spp_taxonconcept() - close #42
